### 2023-11-15 #6061 @editorialbot
[PRE REVIEW]: Spelunker: A quick-look Python pipeline for JWST NIRISS FGS Guidestar Data
**Submitting author:** @GalagaBits (<a href="http://orcid.org/0009-0006-6758-4751">Derod Deal</a>)
**Repository:** https://github.com/GalagaBits/JWST-FGS-Spelunker
**Branch with paper.md** (empty if default branch):
**Version:** v1.02
