### 2023-12-30 b0d2cc4
new save function and methods
