+++ src/spelunker.py
@@ -151,9 +151,23 @@ class load:
 
         self.fontsize = 14
         
+        
         if pid != 'None':
             self.download(pid, obs_num=obs_num, visit=visit, visit_group=visit_group, parallel_sequence_id=parallel_sequence_id,
                           activity_number=activity_number, exposure_number=exposure_number, dir_seg=dir_seg, guider=guider, calib_level=calib_level, save=save, token = token)
+            
+            self.pid = pid
+
+            if obs_num == 'None':
+                self.obs_num = None
+            else:
+                self.obs_num = obs_num
+
+            if visit == 'None':
+                self.visit = None
+            else:
+                self.visit = visit
+            
 
     def stitcher(self, fg_timeseries):
         '''
@@ -509,7 +523,7 @@ class load:
         for i in fg_table['sliced_directory']:
             visit_group_col.append(int(i[14:16]))
             parallel_sequence_id_col.append(int(i[16]))
-            activity_num.append(int(i[17:19]))
+            activity_num.append(str(i[17:19]))
             exposure_number_col.append(int(i[20:25]))
 
             if 'seg' in i:
@@ -550,7 +564,7 @@ class load:
                 fg_table = fg_table[mask5]
 
         if activity_number != 'None':
-                mask6 = fg_table['activity_number'] == int(activity_number)
+                mask6 = fg_table['activity_number'] == str(activity_number)
                 fg_table = fg_table[mask6]
 
         if exposure_number != 'None':
@@ -735,31 +749,181 @@ class load:
         return master_table
     
 
-    def readfile(self, filename):
+    def readfile(self, pid, obs_num='None', visit='None', visit_group='None', parallel_sequence_id='None', 
+                 activity_number='None', exposure_number='None', dir_seg='None', guider='None',):
         '''
         This function opens a singular FG-GS guidestar FITS and writes attributes for ``self.fg_array``, ``self.fg_time``, and ``self.flux``. Note that if a program ID
         is previously loaded using ``load``, the attributes already created will be overwritten. Currently, ``readfile`` does not create ``fg_table`` and ``object_properties``
         attributes.
         '''
 
-        fg2 = datamodels.open(filename)
+        pid = str(pid)
 
-        all_times = np.linspace(fg2.meta.guidestar.data_start, 
-                                fg2.meta.guidestar.data_end, 
-                                fg2.data.shape[0])
+        self.pid = pid
+
+        if obs_num == 'None':
+            obs_num = '***'
+        else:
+            obs_num = str(obs_num).zfill(3)
+
+            self.obs_num = obs_num
+
+        if visit == 'None':
+            visit = '***'
+        else:
+            visit = str(visit).zfill(3)  
+
+            self.visit = visit  
         
-        all_fluxes = np.nansum(fg2.data, axis = (1,2))
+        fg_raw = sorted(glob.glob(self.directory+'/mastDownload/JWST/'+'**/jw0'+pid+obs_num+visit+'**_gs-fg_**cal.fits'))
+
+        if len(fg_raw) == 0:
+            print('\t Could not find any files with these parameters.')
+            raise Exception('No files were read for program '+str(pid))
+
+        print(fg_raw)
+        fg = []
+        sliced_directory = []
+
+        for i in fg_raw:
+                fg.append(i.rsplit('/')[-1])
+                sliced_directory.append(i.split('/')[-2])
+
+        fg_table = Table()
+
+        fg_table['filenames'] = fg
+        print(fg_table['filenames'])
+        fg_table['sliced_directory'] = sliced_directory
+
+        obs_num_col = []
+        visit_col = []
+
+        visit_group_col = []
+        parallel_sequence_id_col = []
+        activity_num = []
+        exposure_number_col = []
+        dir_segment_col = []
+        guider_col = []
+
[…]
