### 2024-01-10 #6202 @editorialbot
[REVIEW]: Spelunker: A quick-look Python pipeline for JWST NIRISS FGS Guidestar Data
**Submitting author:** @GalagaBits (<a href="http://orcid.org/0009-0006-6758-4751">Derod Deal</a>)
**Repository:** https://github.com/GalagaBits/JWST-FGS-Spelunker
**Branch with paper.md** (empty if default branch):
**Version:** v.1.1.17
