### 2024-01-29 #6202 @taylorbell57
@ivastar, I just wanted to make it clear that I have an insigificant "potential conflict of interest" with the second-author Néstor Espinoza. The list of papers where Néstor and I are coauthors are: [1](https://ui.adsabs.harvard.edu/abs/2023Natur.614..649J/abstract), [2](https://ui.adsabs.harvard.edu/abs/2023Natur.614..653A/abstract), [3](https://ui.adsabs.harvard.edu/abs/2023Natur.614..664A/abstract), [4](https://ui.adsabs.harvard.edu/abs/2023Natur.614..670F/abstract), [5](https://ui.adsabs.harvard.edu/abs/2023PASP..135c8002B/abstract), [6](https://ui.adsabs.harvard.edu/abs/2023Natur.620..292C/abstract). For all of papers 1-4 and 6, I was a minor coauthor and didn't actually collaborate with Nestor in any direct way. For paper 5, I contributed one of the analyses to the paper but still did not actually collaborate with Néstor in any direct way.

I do not believe that this will in any way interefere with my ability to impartially review the code or paper, I just wanted to make this clear in advance in compliance with the Conflict of Interest Policy.
