### 2024-01-29 #6202 @taylorbell57
@GalagaBits and @nespinoza, can you clarify what version number we are supposed to be reviewing? On GitHub the lastest release/tag that I see is [v.1.0.2](https://github.com/GalagaBits/JWST-FGS-Spelunker/releases/tag/v.1.0.2), but `pip` installs v1.1.0 from PyPI (which also has v1.0.3 and v1.0.4 which are missing from GitHub).
