### 2024-01-29 #6202 @taylorbell57
Also, note that the images are missing from the PyPI posting at https://pypi.org/project/spelunker/#description
