### 2024-01-29 GalagaBits/JWST-FGS-Spelunker#20 @taylorbell57
Python version upper limit
When trying to install Spelunker with Python 3.12.1, I got the following error:

```
> pip install spelunker
Collecting spelunker
  Downloading spelunker-1.1.0-py3-none-any.whl.metadata (5.0 kB)
Collecting numpy (from spelunker)
  Downloading numpy-1.26.3-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (61 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 61.2/61.2 kB 2.8 MB/s eta 0:00:00
Collecting scipy (from spelunker)
  Downloading scipy-1.12.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (60 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 60.4/60.4 kB 6.4 MB/s eta 0:00:00
Collecting pandas (from spelunker)
  Downloading pandas-2.2.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (19 kB)
Collecting jwst (from spelunker)
  Downloading jwst-1.12.5.tar.gz (18.5 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting astroquery (from spelunker)
  Using cached astroquery-0.4.6-py3-none-any.whl (4.5 MB)
Collecting astropy (from spelunker)
  Downloading astropy-6.0.0-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (9.8 kB)
Collecting astroplan (from spelunker)
  Downloading astroplan-0.9.1.tar.gz (140 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 141.0/141.0 kB 9.7 MB/s eta 0:00:00
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
Collecting matplotlib (from spelunker)
  Downloading matplotlib-3.8.2-cp312-cp312-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.8 kB)
INFO: pip is looking at multiple versions of spelunker to determine which version is compatible with other requirements. This could take a while.
Collecting spelunker
  Downloading spelunker-1.0.4-py3-none-any.whl.metadata (5.0 kB)
  Downloading spelunker-1.0.3-py3-none-any.whl.metadata (5.0 kB)
  Downloading spelunker-1.0.2-py3-none-any.whl.metadata (4.9 kB)
  Downloading spelunker-1.0.1-py3-none-any.whl.metadata (4.9 kB)
  Downloading spelunker-1.0.0-py3-none-any.whl.metadata (4.8 kB)
  Downloading spelunker-0.5.5-py3-none-any.whl.metadata (4.4 kB)
  Downloading spelunker-0.5.4-py3-none-any.whl.metadata (4.3 kB)
INFO: pip is still looking at multiple versions of spelunker to determine which version is compatible with other requirements. This could take a while.
ERROR: Cannot install spelunker==0.5.4, spelunker==0.5.5, spelunker==1.0.0, spelunker==1.0.1, spelunker==1.0.2, spelunker==1.0.3, spelunker==1.0.4 and spelunker==1.1.0 because these package versions have conflicting dependencies.

The conflict is caused by:
    spelunker 1.1.0 depends on ray
    spelunker 1.0.4 depends on ray
    spelunker 1.0.3 depends on ray
    spelunker 1.0.2 depends on ray
    spelunker 1.0.1 depends on ray
    spelunker 1.0.0 depends on ray
    spelunker 0.5.5 depends on ray
    spelunker 0.5.4 depends on ray

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
```

It appears your `ray` dependency is incompatible with Python 3.12.1. However, when I downgraded to Python 3.11.0, the package installed successfully. I recommend you note this upper-limit on the Python version in your documentation and potentially in your setup.py file.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
