### 2024-01-29 GalagaBits/JWST-FGS-Spelunker#21 @taylorbell57
Discrepancies in first figure of README.md
This [linked code](https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/README.md?plain=1#L37C1-L47C5) in the README.md does not exactly produce the subsequent figure shown in the file. I strongly recommend you adjust the code to include everything needed to reproduce the figure (including any custom rcParams settings) so that new users have a quick and easy way to validate their installation with this realistic use case.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
