### 2024-01-29 GalagaBits/JWST-FGS-Spelunker#22 @taylorbell57
Clearing out outdated files and tidying the folder structure
The [src/dashboard.py](https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/dashboard.py#L1-L19) file appears to be entirely commented out. Do you intend for it to still be present, and if so why is it commented out? You also have a [temp.rst](https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/temp.rst) file which I presume is no longer needed (given the separate docs folder, and given the "temp" name).

Additionally, you should not have [`__pycache__`](https://github.com/GalagaBits/JWST-FGS-Spelunker/tree/b0d2cc40c9d8f9872ba1c160bb43765176453de1/__pycache__) or [`.DS_Store`](https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/.DS_Store) folders/files present in your repository (and the latter can actually be dangerous to store, from my understanding). Please add these to your `.gitignore` file and delete them from your repo.

Finally, I also recommend moving your JOSS paper files into a directory (calling it `paper` should still work with JOSS' automations I believe) to try to keep the main directory cleaner.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
