### 2024-01-29 GalagaBits/JWST-FGS-Spelunker#23 @taylorbell57
Missing community guidelines
As required by JOSS, there should be

> clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This can cleanly be done by completing the Code of conduct section at https://github.com/GalagaBits/JWST-FGS-Spelunker/community. JOSS' own code of conduct may be a good starting point: https://github.com/openjournals/joss/blob/main/CODE_OF_CONDUCT.md. The statement of

> The pipline is currently under development on [GitHub](https://github.com/GalagaBits/JWST-FGS-Spelunker). For any bugs or requests, send us an email or [open an issue on GitHub](https://github.com/GalagaBits/JWST-FGS-Spelunker/issues).

on your readthedocs page may suffice for JOSS' purposes, but adding something like that to your README would likely be a good step and adding a code of conduct would be ideal (in my opinion).

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
