### 2024-01-29 GalagaBits/JWST-FGS-Spelunker#24 @taylorbell57
Testing script is currently failing for v1.1.0
The tests are currently failing for v1.1.0 with Python 3.11.0 (see traceback below).

```
> python general_test.py
Warning: jwstuser is not installed. mnemonics() will not work.

 >> Testing spelunker version  1.1.0
Current working directory for spelunker: /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs

Connecting with astroquery...
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159061621_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159061621_cal.fits ...
|==================================================================================| 2.5M/2.5M (100.00%)         0s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159062424_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159062424_cal.fits ...
|==================================================================================| 7.5M/7.5M (100.00%)         1s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg001_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg001_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         5s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg002_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg002_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         6s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg003_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg003_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         5s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg004_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg004_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         5s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg005_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg005_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         5s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg006_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg006_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         6s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg007_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg007_cal.fits ...
|==================================================================================|  51M/ 51M (100.00%)         5s
Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:JWST/product/jw01541001001_gs-fg_2022159132423-seg008_cal.fits to /home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/spelunker_outputs/mastDownload/JWST/jw01541001001_02101_00004-seg001_nis/jw01541001001_gs-fg_2022159132423-seg008_cal.fits ...
|==================================================================================|  35M/ 35M (100.00%)         3s
2024-01-29 14:29:28,303 INFO worker.py:1724 -- Started a local Ray instance.

 >> Results:

         - Downloading & storage of data (spk.load)........PASS
         - 2D gaussian fits (spk.gauss2d_fit)........PASS
         - Guidestar plot (spk.guidestar_plot)........FAIL

         - Final assessment: some tests failed. Here is a summary of those, along with errors raised:

                 1) Guidestar plot (spk.guidestar_plot) failed with a IndexError :

Traceback (most recent call last):
  File "/home/taylor/Refereeing/Spelunker/JWST-FGS-Spelunker/testing/general_test.py", line 72, in test_guidestarplot
    spk.guidestar_plot()
  File "/home/taylor/miniconda3/envs/spelunker/lib/python3.11/site-packages/spelunker/spelunker.py", line 1483, in guidestar_plot
    ax, hdu = plot_finder_image(target, survey='DSS', fov_radius=fov_radius,)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/taylor/miniconda3/envs/spelunker/lib/python3.11/site-packages/astropy/units/decorators.py", line 313, in wrapper
    return_ = wrapped_function(*func_args, **func_kwargs)
              ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/taylor/miniconda3/envs/spelunker/lib/python3.11/site-packages/astroplan/plots/finder.py", line 88, in plot_finder_image
    hdu = SkyView.get_images(position=position, coordinates=coordinates,
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
IndexError: list index out of range

 >> Some spelunker tests failed. Please check errors above.
```

Additionally, these tests should ideally be run automatically to ensure that they are run and to save you devs some work and headaches. GitHub offers automated workflows that can be run on GitHub servers in various circumstances (e.g. for every PR). You could look at [this file](https://github.com/kevin218/Eureka/blob/main/.github/workflows/pytest_testing.yml) I helped make for Eureka! for inspiration.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
