### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#25 @evamariaa
Missing documentation on jwstuser when using mnemonics()
I get this warning every time I load spelunker:
`Warning: jwstuser is not installed. mnemonics() will not work.`

While jwstuser is not part of spelunker I wasn't able to find any information in the documentation and online. I assume this will be a very common warning message so I recommend adding some more details on that or link to the relevant pages.

This issue was made as part of the review for openjournals/joss-reviews/issues/6202.
