### 2024-01-30 #6202 @evamariaa
Minor comment about the documentation @GalagaBits: It'd be useful to have the link to the documentation on the right 'About' section in the GitHub repository. It took me a minute or two to find the link all the way at the bottom of the page.
