### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#26 @evamariaa
Issue with astroquery when calling cycle 2 JWST proposals with spelunker.load
I tried loading spelunker with some cycle 2 JWST proposal IDs (that have definitely been completed) like this:
`spk = spelunker.load(pid=4098, obs_num='13', visit='1', save=True)`

but it gives an error message:
```
WARNING: NoResultsWarning: Query returned no results. [astroquery.mast.discovery_portal]
2024-01-30 13:39:17,619 - stpipe - WARNING - NoResultsWarning: Query returned no results.
```
Was there a recent update that affected the newer JWST observations?

This issue was made as part of the review for openjournals/joss-reviews/issues/6202
