### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#27 @taylorbell57
jwstuser installation missing from installation.rst docs
The `jwstuser` installation instructions are missing from the `installation.rst` docs file. This is described in your README.md file, but should also be described on your installation instructions page

https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/docs/installation.rst?plain=1#L1-L22

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
