### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#21 @taylorbell57
Similarly, the code in the documentation page does not reproduce the shown figure:

https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/docs/simple_quickstart.rst?plain=1#L36-L45
