### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#28 @taylorbell57
Need added text to clarify second figure of README.md
You should add a sentence or two describing what the Guidestar positions figure shows, as it is not immediately clear to me let alone someone new to JWST data. Describe what the symbols mean and tersely describe why JWST would be changing between different guide stars. Any guide stars are also presumably covered by your "x" symbols - perhaps an additional figure (or more than one) or subfigure showing cutouts around the guide stars would be helpful for some use cases?

This terse description should be added to the README.md file as well as the docs/simple_quickstart.rst file

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
