### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#28 @taylorbell57
Similarly, the third figure is not immediately understandable and could use at least a simple description of what is being shown
