### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
FYI, this is related to #27

I recommend at least adding `jwstuser` as an optional installation dependency in the `setup.py` file so that one can do `pip install spelunker[jwstuser]` or something like that to install both together. However, if the additional burden of always installing `jwstuser` is fairly minimal, then I don't see why it couldn't just be one of your required dependencies to save most folks from getting that warning every single time they load the package (especially since such warnings can sometimes get repeated when using parallelization).
