### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#29 @taylorbell57
Feedback on fgs-spelunker-and-tsos.rst documentation
Below is some feedback on your (very helpful!) fgs-spelunker-and-tsos.rst documentation page. I've made this a check list so that it's easy for me to cross things off after you've made some updates.

- [x] The program version on your readthedocs page is currently set to `Program version: ver. 0.5.3` which doesn't exist on GitHub and is greatly out-of-date with the pip installed version 1.1.0. The page should either be remade with the newer code version or the code version should be updated on the page.
- [x] targetted -> targeted
- [x] accreeting -> accreting
- [x] HAT-P-14~b -> HAT-P-14 b
- [x] You should briefly describe where the `'hp14_lightcurve.dat'` and `'w39_lightcurve.dat'` files comes from (citing the source paper and software as appropriate) and explain what package(s) the reader could use to make such a file themselves.
- [x] Adding `plt.show()` at the end of blocks of plotting code will avoid the printing of values like `(0.0, 6.09976451843977)` to the console
- [x] The transit appears to start closer to 2.75 hours after the start, not 2.5
- [x] It would be useful to briefly describe where someone might find the `pid`, `obs_num`, and `visit` values for an observation they want to analyze. Newcomers may not immediately be familiar with these values.
- [x] `The raw photometry from the guidestar varies by a lot more than our target star.` I assume much of this variation is just photon noise given the very short exposure time - if this is the case, it would be useful to briefly describe this to remind the user that the cadence varies between the science instrument and the FGS and to explain why temporal binning would be helpful.
- [x] Rather than making a custom `bin_fgs_to_science` function from scratch, wouldn't it be better to make use of a spelunker function to get users familiar with more of the utilities you have built?
- [x] The shown outputs from running `spk.gauss2d_fit(ncpus=4)` include a large number of error messages. Ideally these should be resolved if possible, they captured and hidden if unimportant, or at a minimum they should be explained in the tutorial so that folks know that these kinds of errors are not concerning. Additionally, there should likely also be either a `mute` or `verbose` option added so that the printed table does not have to be output to the console.
- [x] `..."appears to be a small tilt event."` You should briefly explain what a tilt event is and add a link to a relevant JDox page
- [x] `"Best of all, we can see the event at any resolution we want thanks to it."` Probably best to remind the reader again here that the FGS data have a finer time sampling which is what allows this higher temporal resolution (also likely worth stating temporal resolution to differentiate from spectral resolution)
- [x] Rather than making a new `bin_data` function, you should use your built-in `spelunker.bin_data` function
- [x] `"Interestingly, in this case, the Y-standard deviation samples the event even better."` This is not obvious from the plots. You might try inverting the secondary y-axis for sigma_y and/or plotting both sigma_x and sigma_y after some kind of normalization
- [x] Finally, it would be worth adding a concluding sentence or few that would explain where someone might go from here. Something about how one could use these FGS data as covariates when fitting their lightcurves to remove noise would be useful (e.g. linearly decorrelate against FGS y-pixel position).

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
