### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#30 @taylorbell57
Unusual optimizing_photometry.rst file
The `optimizing_photometry.rst` file ends up being nested under https://jwst-fgs-spelunker.readthedocs.io/en/latest/timeseries_observations.html which feels a very unusual spot for it since it is an API document rather than a tutorial like `fgs-spelunker-and-tsos.rst`. I would recommend just moving the documentation of `optimizing_photometry` to the API page.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
