### 2024-01-30 GalagaBits/JWST-FGS-Spelunker#31 @taylorbell57
Additional textual explanation needed for example notebooks
The example notebooks need some additional textual explanation to be understandable and useful to a new user. Please also check for spelling mistakes throughout (e.g. chuck->chunk, obserbed->observed). More importantly,

**`guidestar_targets.ipynb`:**
This notebook in particular needs a fair bit of explanation to be useful. There is no description of what the observations were doing, why the guide star changed so many times, and what utility there is in the demonstrated figures (what do they show, how could they be used). You've also written that you're going to use PID 2079 but actually use PID 1534. You also included an actual `mast_api_token` in Cell [6] which is a really bad idea (thankfully the API token is 5 months old now and will have already expired). You should instead be loading this from an external file/variable using something like `spk.mast_api_token = os.environ.get('MAST_API_TOKEN')` to ensure this never happens.

**`pixel_centroid_mnemonics.ipynb`:**
Some context at the top of the notebook describing what the notebook is going to teach folks would be very helpful. Also briefly explaining the observations of PID 2079 would help set the context some. I'm not really sure what is meant by `"We can see that there are chucks of data points that largely changes in flux for each chunk."` The outputs of `In [24]` and `In [15]` seem to show that the time-axis may not be sorted, resulting in some instances where there are two lines for a given x value (because pyplot by default is just connecting the dots in the order they were provided); this may hint at a more concerning underlying issue with the ordering of the values gathered by spelunker. Either way, its probably best to use points instead of connected lines to more clearly demonstrate the breaks between observations.

**`pixel_centroid_mnemonics.rst`:**
There seems no need to me to have an rst file in addition to the notebook, and all of the figures that went along with the notebook are broken in this rst version.

Also, note that some of the outputs (in particular the tables) and even some of the inputs extend too far on your readthedocs page, making them unreadable at times.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
