### 2024-01-31 GalagaBits/JWST-FGS-Spelunker#32 @taylorbell57
Readthedocs pages poorly render on smaller screens
At present, your documentation at https://jwst-fgs-spelunker.readthedocs.io/ is poorly optimized for smaller screens (e.g. laptops and especially phones), requiring lots of horizontal scrolling. This must just be caused by the particular template that you have chosen since most other readthedocs pages that I've seen adapt very well to differently sized screens (e.g. https://eurekadocs.readthedocs.io/)

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
