### 2024-01-31 GalagaBits/JWST-FGS-Spelunker#33 @taylorbell57
Feedback on docs/user/fgs-spelunker_quickstart.rst
Below is some feedback on your `docs/user/fgs-spelunker_quickstart.rst` documentation page. I've made this a check list so that it's easy for me to cross things off after you've made some updates.

I'll also note that this notebook is particularly useful in explaining many of the detailed operations of the code. You likely want to make this more enticing for people to click and read rather than just calling it "Spelunker Guidestar Tools" which doesn't make it seem as useful as it is as a more in-depth tutorial. Putting this right below the Quickstart would likely also help convince new users to check it out - this far down the list, I thought it'd be some random extra utilities.

- [x] #40
- [x] `Currently, ``spk.readfile()`` does not support the folling
attributes: - ``fg_table``, - ``object_properties``.` Is this a comment meant to be noted outside of the code environment, or is this a warning message that is being output by the code? (I presume the former)
- [x] `"Current supported mnemonics: SA_ZHGAUPST (high-gain antenna), INIS_FWMTRCURR (NIRISS Filter Wheel Motor Current)"`. What barriers are there to accessing additional (arbitrary) mnemonics? It'd be amazing if folks could request any random mnemonic and get a result (I know there'd be several that I'd like to experiment with), but I also understand that you don't want to always grab all mnemonics since that'd be very wasteful.
- [x] `https://jwst-fgs-spelunker.readthedocs.io/en/latest/_images/fgs-spelunker_quickstart_51_1.png` Would it be possible to make this a GIF for the purposes of the readthedocs page so that web users can see what this animation looks like?

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
