### 2024-01-31 GalagaBits/JWST-FGS-Spelunker#34 @taylorbell57
Malfunctioning Python Module Index on readthedocs
The "Python Module Index" found at https://jwst-fgs-spelunker.readthedocs.io/en/latest/py-modindex.html which is accessed through clicking on the MODULES link at the top of the readthedocs page only shows `spelunker`, and when you click on that it just takes you to https://jwst-fgs-spelunker.readthedocs.io/en/latest/optimizing_photometry.html#module-spelunker.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
