### 2024-01-31 GalagaBits/JWST-FGS-Spelunker#35 @taylorbell57
Feedback on API documentation
Below is some feedback on your [API](https://jwst-fgs-spelunker.readthedocs.io/en/latest/api.html) documentation page. I've made this a check list so that it's easy for me to cross things off after you've made some updates.

- [x] Please ensure that all functions have all of their inputs and (if relevant) outputs documented with a description of the variables and their data types. You've done this for many but not all functions. To save you some time, in some places you can just copy-paste the documentation for parameters that are already documented in the `load` function.
- [x] There is a missing space between the parameter names and parameter types in the documentation for basically every function except for `load`
- [x] In `gauss2d_fit`, you should explain briefly what all of the different parameters mean (either in text or by linking to some astropy page or whatever), since a newcomer might be confused by parameters like theta and offset.
- [x] Check for typos throughout (e.g. "speficied"->"specified")
- [x] It seems unusual to me to force the plotting of the mnemonics in the `mnemonics` and `mnemonics_local` functions. Having a `plot` input boolean parameter would seem more appropriate. Additionally, is the `mnemonics_event` attribute set by the `mnemonics` function some kind of a dictionary that will store multiple mnemonics or is it something that'll be overwritten every time the `mnemonics` function is called with a new parameter?
- [ ] There is a confusing mixture of object-oriented programming and functional programming. For example, some functions like `negative_flux_preprocessing` require that data arrays be passed in, while functions like `normalization_flux_preprocessing` seem not to accept inputs and instead work on the internally stored arrays, while yet other functions like `gauss2d_fit` are a hybrid and will either operate on the internally stored arrays or on an array provided as an input. There should be more consistency in what you allow for each function to help the users gain some intuition for how to use your code.
- [x] It is likely also best that the function `periodogram` not force plotting but offer it as a (default on) option.
- [x] How is the binning inside `timeseries_binned_plot` controlled? Why not make parameters like bin size available to the user?
- [x] The documented parameters for `timeseries_list_plot` are not in the same order as they are input into the function.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
