### 2024-01-31 GalagaBits/JWST-FGS-Spelunker#35 @taylorbell57
Also, you should quite likely add documentation for the attributes for your `load` class, so that users know what attributes exist and what they contain
