### 2024-02-01 GalagaBits/JWST-FGS-Spelunker#36 @taylorbell57
Feedback on the code
Below is some feedback on your code itself:

---

The following snippet could be greatly simplified in the following way:
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L287-L296
```
bool_mask = fg_flux >= 0
```

---

The `astropy.table.Table.sort` function operates on the existing table and does not return a new table, so there's no need to capture the returned `None` value (and same goes for every other time you call `sort`)
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L336-L338

---

The code in the following lines is used in all the if/elif/else clauses, so it should be moved before the if statement to avoid repeated code and to make maintaining the code easier:
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L397-L410

---

Rather than using the code in this snippet, you could much more quickly and easily determine the brightest few pixels by doing
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1251-L1273
```
mask = np.zeros_like(frame.flatten())
good_inds = np.argsort(frame.flatten(), axis=None)[-npixels:]
mask[good_inds] = 1
return mask.reshape(frame.shape)
```

---

In the following snippets, the `and` operator is not working the way that you intend it to. For example, in the first snippet instead of determining whether both `fg_flux` and `fg_time` are of type `str`, the code is actually doing `if (type(fg_flux)) and (type(fg_time)==str)` which is definitely not the intended meaning of the code since this will return `True` as long as `type(fg_flux)` is not "falsy" which I don't think it will ever be. These are just the few cases where I noticed you made this mistake, you should look through the code to see if there are any other places where you've done this.
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1320-L1321
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1325
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1392-L1393
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1397

---

Why is the MAST API token sometimes passed in as a function parameter but assumed to be an attribute for the following snippet? I recommend sticking with one consistent interface, either only offering it as a parameter for all functions, only offering it as an attribute for all functions, or offering both options for all functions. Additionally, I believe that doing `val == None` is risky and discouraged; I believe you should instead be doing `val is None` throughout your code.
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1536

---

For the following two functions, it would make much more sense to me to allow plotting if an input parameter was set to `True` but not force people to plot just to access the parameter. Additionally, I would think the retrieved values should be returned from the function; that way, someone doesn't pass in `"SA_ZHGAUPST"` and then not know how to access the values without reading the docs and seeing that they have to then access `spk.mnemonics_event_hga`
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1497
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L1615

---

Why not save the estimated parameters (like centroids and PSF widths) if `self.quickfit_results` is not `None`?
https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/b0d2cc40c9d8f9872ba1c160bb43765176453de1/src/spelunker.py#L2015-L2017

---

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
