### 2024-02-01 GalagaBits/JWST-FGS-Spelunker#22 @taylorbell57
I believe the files in the `observations` folder probably do not need to be hosted on this main GitHub repo anymore
