### 2024-02-02 GalagaBits/JWST-FGS-Spelunker#33 @taylorbell57
Additionally, running through the notebook myself results in different numerical results (and they are consistently different, so it's not just being caused by different random number seeds or whatever). I've attached my completed notebook (zipped so GitHub doesn't complain) so you can see that its results differ from those in your uploaded notebook at `notebooks/fgs-spelunker_quickstart.ipynb`.

[fgs-spelunker_quickstart_tjb.zip](https://github.com/GalagaBits/JWST-FGS-Spelunker/files/14132488/fgs-spelunker_quickstart_tjb.zip)
