### 2024-02-02 #6202 @taylorbell57
Alright, I've now finished reading through the paper, docs, and code and have opened several issues (linked above) in addition to the few small comments I made above which will hopefully make them more user friendly, robust, and easy to maintain. That wraps up my first round referee report.

Overall, great work on this @GalagaBits and @nespinoza - very excited to experiment with this in my own research!
