### 2024-02-05 GalagaBits/JWST-FGS-Spelunker#37 @evamariaa
Feedback on the paper summary text
I'd recommend adding a little bit more about the exact applications and functionalities of this package in the paper itself so that people who see this on the arxiv have a bit more of an idea whether this could be useful for their science e.g.
- instead of 'many possible analyses relevant to existing science programs' include some examples,
- add a figure or two from the documentation about how this data looks like
- a description of the 2D Gaussian fit e.g. regarding to identifying tilt events.

This is of course up to you!

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
