### 2024-02-05 #6202 @evamariaa
I think I can also finish up my first review here, the majority of things on my list are covered in the issues Taylor's opened as well and if not I linked the issues I opened above. I haven't had the chance yet to reproduce all example notebooks and go through every bit of documentation in great detail but I see that Taylor has done an amazing job in doing that so I'm happy for this to move forward!
It's a great package @GalagaBits and @nespinoza - well done!
