### 2024-02-20 GalagaBits/JWST-FGS-Spelunker#26 @GalagaBits
I am not completely sure why that error occurs. I think this is a astroquery issue, but I need to look more into it to be sure. However, when I run it, the line does run and I am able to print out a timeseries using `spk.timeseries_binned_plot()`. Are you able to get the same result?
![image](https://github.com/GalagaBits/JWST-FGS-Spelunker/assets/34511693/559614e6-d86d-4a1a-8738-7cea4f89ecf0)
