### 2024-03-05 GalagaBits/JWST-FGS-Spelunker#26 @GalagaBits
Thanks for the reply. I have to ask a few more question to investigate this issue, then I'll fix it to the best of my ability.

- Does the error `WARNING: NoResultsWarning: Query returned no results. [astroquery.mast.discovery_portal]
2024-01-30 13:39:17,619 - stpipe - WARNING - NoResultsWarning: Query returned no results.` occur with every program ID in Cycle 2 data?

- Are you able to still access features like time series plots and arrays for each Cycle 2 proposal after the error occurs?

- Are there any other errors or issues with Spelunker after receiving the error?

Thank you for using Spelunker and again I will fix this to the best of my ability.
