### 2024-03-05 d950a1d
removed dashboard and added new readthedocs theme
