### 2024-03-17 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
Okay, I will add `git+https://github.com/spacetelescope/jwstuser.git` to the dependencies list
