+++ setup.py
@@ -20,8 +20,8 @@ setup(name='spelunker',
       license='MIT',
       packages=['spelunker'],
       package_dir={'spelunker': 'src'},
-      requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray'],
-      install_requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray', 'wheel'],
+      requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray', 'git+https://github.com/spacetelescope/jwstuser.git'],
+      install_requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray', 'git+https://github.com/spacetelescope/jwstuser.git', 'wheel'],
       keywords='guidestars',
       python_requires='>=3.10',
       zip_safe=False)
