### 2024-03-17 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
@taylorbell57, I have tried added `jwstuser @ git+https://github.com/spacetelescope/jwstuser@main#egg=jwstuser` to `install_requires` in my `setup.py` file. However, I get the error

```
ERROR    HTTPError: 400 Bad Request from
         https://upload.pypi.org/legacy/
         Invalid value for requires_dist. Error: Can't
         have direct dependency: 'jwstuser @
         git+https://github.com/spacetelescope/jwstuser@m
         ain#egg=jwstuser'
```

when trying to upload to PyPI. I'm not sure if there's currently a way to add direct dependencies into PyPI right now
