+++ docs/installation.rst
@@ -20,3 +20,8 @@ The following packages are required:
 * astroplan (optional; for use with :func:`spelunker.load.guidestar_plot`)
 
 Additonally, you can refer to the `GitHub repository <https://github.com/GalagaBits/JWST-FGS-Spelunker/>`_ for the source code.
+
+### JWST User Toolkit (`jwstuser`)
+
+To use the `mnemonics` engineering telemetry features of Spelunker, the python toolkit `jwstuser` will need to be installed as well.
+Follow the instructions on `https://github.com/spacetelescope/jwstuser <https://github.com/spacetelescope/jwstuser>`_ to install `jwstuser`.
