+++ src/spelunker.py
@@ -2074,4 +2074,4 @@ class load:
                                                                                                                     self.pgram_results['power_offset'].value[j],
                                                                                                                     ))
 
-                fout.close() break
+                fout.close()
