+++ docs/conf.py
@@ -53,7 +53,7 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 #
 html_logo = "spklogo3.png"
 
-html_theme = 'latest'
+html_theme = 'master'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
