+++ docs/conf.py
@@ -28,7 +28,7 @@ author = 'Derod Deal & Nestor Espinoza'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc'
+    'sphinx_rtd_theme'
 ]
 
 # Add any paths that contain templates here, relative to this directory.
