+++ docs/installation.rst
@@ -1,5 +1,5 @@
 Installation
-============
+############
 
 ``spelunker`` can be installed using ``pip``.
 
@@ -21,7 +21,8 @@ The following packages are required:
 
 Additonally, you can refer to the `GitHub repository <https://github.com/GalagaBits/JWST-FGS-Spelunker/>`_ for the source code.
 
-### JWST User Toolkit (`jwstuser`)
+JWST User Toolkit (`jwstuser`)
+******************************
 
 To use the `mnemonics` engineering telemetry features of Spelunker, the python toolkit `jwstuser` will need to be installed as well.
 Follow the instructions on `https://github.com/spacetelescope/jwstuser <https://github.com/spacetelescope/jwstuser>`_ to install `jwstuser`.
