### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#27 @GalagaBits
Okay, I have added `jwstuser` instructions on the `installation.rst` page.
