### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#36 @GalagaBits
@taylorbell57, can you explain this section of your feedback:

```
The astropy.table.Table.sort function operates on the existing table and does not return a new table, so there's no need to capture the returned None value (and same goes for every other time you call sort)
```

I do not think I understand the issue
