### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#32 @GalagaBits
Okay, I have changed the readthedocs template to a more modern and mobile friendly theme!
