+++ docs/index.rst
@@ -2,8 +2,9 @@ Spelunker
 =========
 
 .. image:: spelunker.png
+    :width: 400pt
     :align: center
-    :scale: 7%
+    
 
 ``spelunker`` is a package that assists and finds technical anomalies on the JWST and stellar properties from guidestars. This pipeline includes tools for
 conveniently handling guidestar data products and provides methods for investigating its potential in scientific and technical observations.
