### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#20 @GalagaBits
Okay, I have updated the installation documentation to account for this, as well as setting the python version in the setup.py file to Python 3.10!
