### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
For now, I have added extra documentation on the installation page guiding users to the `jwstuser` github per #27 for installation instructions.
