### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#22 @GalagaBits
I have addressed all of these issues! Also, dashboard.py was a older script that was apart of some sort of user interface for Spelunker. For now, I have removed it since it is not yet apart of Spelunker.
