### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#22 @GalagaBits
I think moving the paper.md file to another directory might break the JOSS automations. Please let me know if that is the case in the future.
