### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#36 @taylorbell57
As you can see from their documentation page [here](https://docs.astropy.org/en/stable/api/astropy.table.Table.html#astropy.table.Table.sort), the `sort` function doesn't return anything and instead edits the data "in place". Doing `sort_table = table.sort(['fg_time'])` makes it seem like `sort_table` would contain the sorted table and should be the one returned and not `table` as is currently done, but in reality `sort_table` ends up being set to `None`. That line should just be replaced with just `table.sort(['fg_time'])` to not confuse future users or even yourself in future months/years
