### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#20 @taylorbell57
Note that you've set a lower-limit of 3.10, but you haven't set an upper limit. The line should instead be `python_requires='>=3.10,<3.12'`, and assuming that Python 3.9 or lower would work (which I suspect it may but haven't tested) you could simply just do `python_requires='<3.12'`
