### 2024-03-18 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
@GalagaBits, I took a look at your setup.py file, and I was able to get the following to work. Remove the `requires` line (not clear to me that is needed and it seems to be the source of some issues), remove the `import numpy` line (since this requires users to first manually install numpy which isn't required since numpy isn't used during setup), and then set the `install_requires` line to

```
install_requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray', 'jwstuser@git+https://github.com/spacetelescope/jwstuser.git', 'wheel'],
```
