### 2024-03-18 8c5a596
updates to object_properties_func
