### 2024-03-18 dfae453
updates to object_properties_func
