### 2024-03-18 1abfc4b
updates to object_properties_func
