+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.4] - 2023-03-18 (Current release)
+## [1.1.5] - 2023-03-18 (Current release)
+
+- Updates to `object_properties_func`.
+
+## [1.1.4] - 2023-03-18
 
 - Updates to `object_properties_func`.
 
