### 2024-03-19 cb0964e
updates to spelunker.load.table
