### 2024-03-19 f6912c7
updates to spelunker.load.table
