+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed an issue where the searching the guidestar catalog with certain guidestars IDs will not be found, thus breaking `spelunker.load`. 
 
-## [1.1.5] - 2023-03-18 (Current release)
+## [1.1.5] - 2023-03-18
 
 - Updates to `object_properties_func`.
 
