### 2024-03-19 #6202 @GalagaBits
@ivastar I have been addressing the comments as issues on https://github.com/GalagaBits/JWST-FGS-Spelunker/issues. I plan to complete the rest of the comments by the end of the month.
