+++ setup.py
@@ -1,5 +1,4 @@
 import re
-import numpy
 from setuptools import setup, Extension, find_packages
 
 VERSIONFILE='src/_version.py'
@@ -20,8 +19,7 @@ setup(name='spelunker',
       license='MIT',
       packages=['spelunker'],
       package_dir={'spelunker': 'src'},
-      requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray',],
-      install_requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray', 'wheel'],
+      install_requires=['numpy','scipy', 'pandas', 'jwst', 'astroquery', 'astropy','astroplan','matplotlib', 'ray', 'jwstuser@git+https://github.com/spacetelescope/jwstuser.git', 'wheel'],
       keywords='guidestars',
       python_requires='>=3.10',
       zip_safe=False)
