+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.6] - 2023-03-18 (Current release)
+## [1.1.7] - 2023-03-19 (Current release)
+
+- Added `jwstuser` to `setup.py`.
+
+## [1.1.6] - 2023-03-18
 
 - Fixed an issue where the searching the guidestar catalog with certain guidestars IDs will not be found, thus breaking `spelunker.load`. 
 
