### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
When I follow this, I still get the exact error

```
100% ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 30.5/30.5 kB • 00:00 • 203.2 MB/s
WARNING  Error during upload. Retry with the --verbose option for more details.
ERROR    HTTPError: 400 Bad Request from https://upload.pypi.org/legacy/
         Invalid value for requires_dist. Error: Can't have direct dependency: 'jwstuser @ git+https://github.com/spacetelescope/jwstuser.git'
```
