### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
I use the following lines to package the library if it helps:

`python setup.py sdist bdist_wheel`

`twine upload --repository pypi dist/*`
