### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
Aha, it is because I used `pip install .` instead of `python setup.py install`. The pip install method is now the recommended method, while the python setup.py install method is an old and now discouraged way of installing python packages
