### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
Ah, so if I use `pip install . sdist bdist_wheel`, it should work?
