+++ docs/user/fgs-spelunker-and-tsos.rst
@@ -306,404 +306,6 @@ bit patient:
     spk.gauss2d_fit(ncpus=4)
 
 
-.. parsed-literal::
-
-    2023-07-28 15:06:29,069	ERROR services.py:1207 -- Failed to start the dashboard , return code 1
-    2023-07-28 15:06:29,072	ERROR services.py:1232 -- Error should be written to 'dashboard.log' or 'dashboard.err'. We are printing the last 20 lines for you. See 'https://docs.ray.io/en/master/ray-observability/ray-logging.html#logging-directory-structure' to find where the log file is.
-    2023-07-28 15:06:29,097	ERROR services.py:1276 -- 
-    The last 20 lines of /tmp/ray/session_2023-07-28_15-06-25_792862_10261/logs/dashboard.log (it contains the error message from the dashboard): 
-    2023-07-28 15:06:28,974	INFO head.py:226 -- Loading DashboardHeadModule: <class 'ray.dashboard.modules.usage_stats.usage_stats_head.UsageStatsHead'>
-    2023-07-28 15:06:28,974	INFO head.py:239 -- Loaded 1 modules. [<ray.dashboard.modules.usage_stats.usage_stats_head.UsageStatsHead object at 0x106db63a0>]
-    2023-07-28 15:06:28,974	INFO head.py:331 -- http server disabled.
-    2023-07-28 15:06:28,974	ERROR head.py:346 -- Failed to connect to GCS. Please check `gcs_server.out` for more details.
-    2023-07-28 15:06:28,987	ERROR dashboard.py:259 -- The dashboard on node rakisduam.stsci.edu failed with the following error:
-    Traceback (most recent call last):
-      File "/Users/nespinoza/opt/anaconda3/envs/newen/lib/python3.9/site-packages/ray/dashboard/dashboard.py", line 248, in <module>
-        loop.run_until_complete(dashboard.run())
-      File "/Users/nespinoza/opt/anaconda3/envs/newen/lib/python3.9/asyncio/base_events.py", line 647, in run_until_complete
-        return future.result()
-      File "/Users/nespinoza/opt/anaconda3/envs/newen/lib/python3.9/site-packages/ray/dashboard/dashboard.py", line 75, in run
-        await self.dashboard_head.run()
-      File "/Users/nespinoza/opt/anaconda3/envs/newen/lib/python3.9/site-packages/ray/dashboard/head.py", line 346, in run
-        self.gcs_client.internal_kv_put(
-      File "python/ray/_raylet.pyx", line 2132, in ray._raylet._auto_reconnect.wrapper
-      File "python/ray/_raylet.pyx", line 2120, in ray._raylet._auto_reconnect.wrapper
-      File "python/ray/_raylet.pyx", line 2221, in ray._raylet.GcsClient.internal_kv_put
-      File "python/ray/_raylet.pyx", line 410, in ray._raylet.check_status
-    ray.exceptions.RpcError: failed to connect to all addresses
-    2023-07-28 15:06:50,551	INFO worker.py:1621 -- Started a local Ray instance.
-    [2m[33m(raylet)[0m [2023-07-28 15:07:29,521 E 10406 36357164] (raylet) file_system_monitor.cc:111: /tmp/ray/session_2023-07-28_15-06-25_792862_10261 is over 95% full, available space: 26746404864; capacity: 994662584320. Object creation will fail if spilling is required.
[…]
