+++ .github/workflows/general_test.yml
@@ -1,0 +1,28 @@
+name: Spelunker Tests
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  test:
+    name: general_test
+    runs-on: 'ubuntu-latest'
+    env:
+        OS: 'ubuntu-latest'
+        PYTHON: '3.10'
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: pysetup
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.10 
+
+      - name: Install Spelunker
+        run: pip install spelunker  
+
+      - name: Run general_test.py
+        run: python general_test.py  
