+++ .github/workflows/general_test.yml
@@ -8,10 +8,10 @@ on:
 jobs:
   test:
     name: general_test
-    runs-on: 'ubuntu-latest'
+    runs-on: ubuntu-latest
     env:
         OS: 'ubuntu-latest'
-        PYTHON: '3.10'
+        PYTHON: '3.10.8'
     steps:
       - name: Checkout
         uses: actions/checkout@v4
@@ -19,7 +19,7 @@ jobs:
       - name: pysetup
         uses: actions/setup-python@v5
         with:
-          python-version: 3.10 
+          python-version: 3.10.8 
 
       - name: Install Spelunker
         run: pip install spelunker  
