### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#24 @GalagaBits
The latest release fixes this issue! I have also added GitHub Actions for `general_test.py`!
