### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
I personally have never used sdist or bdist_wheel and haven't uploaded anything to pypi, so I recommend you experiment, but I think `pip install . sdist bdist_wheel` should work for your purposes
