### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#26 @GalagaBits
@evamariaa I have submitted an issue to astroquery about this https://github.com/astropy/astroquery/issues/2972.
