+++ docs/optimizing_photometry.rst
@@ -1,8 +1,0 @@
-Optimizing Guidestar Photometry
-===============================
-
-.. module:: spelunker
-
-.. autoclass:: spelunker.load.optimize_photometry
-   :members:
-   :noindex:
