### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#34 @GalagaBits
I have fixed this issue! I removed `optimizing_photometry` since the information was already provided in `API`.
