### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#30 @GalagaBits
I have fixed this issue by removing `optimizing_photometry` since the document contents is already available in `API.rst`.
