### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#28 @GalagaBits
Okay, I have completed this issue, so I closed it. However, @taylorbell57 do you mean I should zoom into places where multiple X's overlap?
