### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#28 @taylorbell57
My point was that there are no obvious stars underneath the X's, presumably because the X-marker is large enough to hide them. I'd meant that it might be helpful to have some kind of subfigure or separate figure showing a small area around the guide stars (e.g., 9x9 pixels or something) just so that users can actually see the guide stars (to look at their morphology or something)
