### 2024-03-19 GalagaBits/JWST-FGS-Spelunker#28 @GalagaBits
Ah, okay, I'll create a new code to achieve that!
