### 2024-03-20 GalagaBits/JWST-FGS-Spelunker#28 @GalagaBits
I made the stars that were covered by the X's more visible by changing the marker shapes and increasing the transparency in v1.1.9.
![image](https://github.com/GalagaBits/JWST-FGS-Spelunker/assets/34511693/960bac24-3765-4b13-a78c-6d4df1cdf0a4)
