+++ docs/installation.rst
@@ -19,6 +19,8 @@ The following packages are required:
 * ray (required for parallel processing of Gaussian fitting)
 * astroplan (optional; for use with :func:`spelunker.load.guidestar_plot`)
 
+For Python 3.12, ``ray`` fails to install with Spelunker when using ``pip install spelunker``. We recommend using Python 3.10 when installing ``ray`` and ``spelunker``.
+
 Additonally, you can refer to the `GitHub repository <https://github.com/GalagaBits/JWST-FGS-Spelunker/>`_ for the source code.
 
 JWST User Toolkit (`jwstuser`)
