### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#26 @GalagaBits
@evamariaa, according to https://github.com/astropy/astroquery/issues/2972, the program ID 4098 and others with this error have not been executed for observations yet. Certain Cycle 1 observations also have this error as well, and that is also due to the observations not yet being executed. Let me know if there is any other questions!
