### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#33 @GalagaBits
Also, @taylorbell57, I have renamed the documentation page to 'Exploring Spelunker tools and features'.
