### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#33 @GalagaBits
I accidently converted it into an issue and deleted it, but for the following comment

>  "The Program IDs that can be loaded are limited to programs without an exclusive access period or are available publicly." Is this still the case when using a MAST API token? If so, that's a really unfortunate limitation since it prohibits people from using the tool for their own data until the data become public (after which time most people will have already published their data). If not, then the ability to use MAST API tokens should be pointed out here.

I meant to say that if you do not have access to a MAST API token, you can only download and use publicly available Program IDs. However, with the `readfile` function, you can use fine guidance files you already have downloaded locally, and with the current version, with no drawbacks. If you do have a MAST API token, you can have access to any program under that token.
