### 2024-03-22 e3802ad
changes to exploring spelunker docs
