+++ docs/user/fgs-spelunker_quickstart.rst
@@ -6,7 +6,7 @@ Exploring Spelunker tools and features
 **Authors:** Derod Deal (dealderod@ufl.edu), Néstor Espinoza
 (nespinoza@stsci.edu) \| **Last update:** July 24, 2023
 
-**Program version:** ver. 1.0.1
+**Program version:** ver. 1.1.9
 
 The JWST telescope carries four different instruments: NIRCam, NIRSpec,
 MIRI and FGS/NIRISS — the latter containing the Fine Guidance Sensor
@@ -40,13 +40,6 @@ Getting started
 To get started with FGS Spelunker, first call ``spelunker.load`` into a
 variable while setting a given Program ID.
 
-.. code:: ipython3
-
-    import os
-    import sys
-    os.chdir('/Users/ddeal/JWST-Treasure-Chest/')
-    sys.path.append('/Users/ddeal/JWST-FGS-Spelunker/JWST-FGS-spk-main/src/')
-
 .. code:: python
 
    import spelunker
@@ -54,7 +47,7 @@ variable while setting a given Program ID.
 
 Calling load without the pid parameter ``spelunker.load()`` will
 initialize ``spelunker`` without downloading any of the files. This is
-useful if you already have timeseries arrays to work with.
+useful if you already have guidestar FITS available locally.
 
 Downloading data
 ----------------
@@ -69,7 +62,7 @@ saved. You can define your own directory by using ``dir=``.
    exclusive access period or are available publicly
    (https://jwst-docs.stsci.edu/accessing-jwst-data/jwst-data-retrieval/data-access-policy#DataAccessPolicy-Exclusiveaccessperiod).
 
-.. code:: ipython3
+.. code:: python
 
     import spelunker
     spk = spelunker.load(dir='/Users/ddeal/JWST-Treasure-Chest/', pid='1534')
@@ -112,7 +105,7 @@ directory and create a 2D array of the guidestar data as well as an
 array of time and a flux timeseries. The same parameters work with
 ``spelunker.load``.
 
-.. code:: ipython3
+.. code:: python
 
     spk2 = spelunker.load(pid=1534, obs_num='2', visit='1', calib_level=2)
     spk2.download(1534, obs_num='2', visit='2', calib_level=2) # This overwrites the object data in spk2 with data from the input parameters
@@ -153,7 +146,7 @@ and flux arrays for all FITS files images under the specified Program
 ID. Use the attributes ``spk.fg_array``, ``spk.fg_time``, and
 ``spk.fg_flux`` to access the arrays.
 
-.. code:: ipython3
+.. code:: python
 
     spk2.fg_array.shape, spk2.fg_time.shape, spk2.fg_flux.shape
 
@@ -172,15 +165,19 @@ parameter, ``spk.download`` will automatically stitch the data from the
 files into an array based on the date and time for each file, along with
 the time and flux arrays.
 
-FGS Spelunker can also handle single fits files already stored locally
+FGS Spelunker can also handle guidestar FITS already stored locally
 by using:
 
 .. code:: python
 
-   spk2.readfile('/Users/ddeal/Spelunker-older/JWST-Treasure-Chest-2023/mastDownload/JWST/jw01534001001_03101_00001_guider1/jw01534001001_gs-fg_2022340000825_cal.fits')
+   spk2 = spelunker.load()
+   spk2.readfile(pid=1534)
 
-   Currently, ``spk.readfile()`` does not support the folling
-   attributes: - ``fg_table``, - ``object_properties``.
+
+Any files under the initialized directory and specified Program ID, observation number, and visit number will be loaded into `spk2`.
+
[…]
