### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#33 @GalagaBits
> https://jwst-fgs-spelunker.readthedocs.io/en/latest/_images/fgs-spelunker_quickstart_51_1.png Would it be possible to make this a GIF for the purposes of the readthedocs page so that web users can see what this animation looks like?

Yes, I just fixed it!
