### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#36 @GalagaBits
@taylorbell57 For this

> In the following snippets, the and operator is not working the way that you intend it to. For example, in the first snippet instead of determining whether both fg_flux and fg_time are of type str, the code is actually doing if (type(fg_flux)) and (type(fg_time)==str) which is definitely not the intended meaning of the code since this will return True as long as type(fg_flux) is not "falsy" which I don't think it will ever be. These are just the few cases where I noticed you made this mistake, you should look through the code to see if there are any other places where you've done this.

What code do you recommend me to do instead?
