### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#26 @evamariaa
It's true that not all observations in programme ID 4098 have been executed but the majority has and the one I queried for definitely has (that's why I do get the FGSS data). So does this mean none of the JWST programmes that have been fully completed will be able to use the astroquery part of spelunker?
