### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#36 @taylorbell57
You need to do the check on each parameter separately. For example, instead of doing
```
if type(fg_flux) and type(fg_time) == str:
```
you would need to do something like
```
if type(fg_flux) == str and type(fg_time) == str:
```
