### 2024-03-22 GalagaBits/JWST-FGS-Spelunker#33 @taylorbell57
> For the third checklist, there is no barrier to access the mnemonics. However, there are thousands of different mnemonics to access and I was not completely sure if all of them will work with the `mnemonics` function. Users can input any mnemonics they want, but now going over it, should I give users the engineering telemetry list so they can experiment with different mnemonics themselves?

Gotcha, you should probably just add a note saying that other mnemonics may work but are not officially supported, and adding a link to a list of parameters or something like [this page](https://outerspace.stsci.edu/display/MASTDOCS/Engineering+Data#EngineeringData-Mnemonics) would be helpful
