### 2024-04-09 GalagaBits/JWST-FGS-Spelunker#36 @GalagaBits
Okay, I have implemented these suggestions in the code and I will release it soon. For now, I think I will leave `mnemonics` to output the plot and have users access it with an attribute for now since I am thinking about rewriting the mnemonics functions entirely to implement your suggestions.

Also, for the token suggestion, the only functions that takes a token as a parameter is the `spelunker.load` class and `download` function. I think it giving a choice between using a parameter or setting the token with `self.mast_api_token` in case users would like to change the token without using the `load` or `download` functions and rewriting their `spk` variables
