+++ src/spelunker.py
@@ -118,7 +118,7 @@ class load:
         '''
         The initalization function for ``load``. This function reads the current directory and initalizes essential attributes.
         '''
-
+        
         self.init_dir = os.getcwd()
 
         created_dir = "spelunker_outputs"
@@ -146,6 +146,7 @@ class load:
 
         self.gaussfit_results = None
         self.quickfit_results = None
+        self.pgram_results = None
 
         self.object_properties = None
 
@@ -324,7 +325,7 @@ class load:
         table['fg_time'] = fg_time
         table['fg_flux'] = fg_flux
 
-        sort_table = table.sort(['fg_time'])
+        table.sort(['fg_time'])
 
         return np.array(table['fg_array']), np.array(table['fg_time']), np.array(table['fg_flux'])
 
@@ -377,6 +378,12 @@ class load:
 
             Observations.login(token=token)
 
+        if token is None:
+
+            if self.mast_api_token is not None:
+
+                Observations.login(token=self.mast_api_token)
+
         self.pid = pid
 
         if obs_num == 'None' and visit != 'None':
@@ -384,23 +391,24 @@ class load:
 
         print("Connecting with astroquery...")
         
+
+        matched_obs = Observations.query_criteria(
+            obs_collection = 'JWST',
+            proposal_id=str(pid),
+        )
+        
+        data_products = [Observations.get_product_list(obs) for obs in matched_obs]
+        data_products = vstack(data_products)
+
+        products = Observations.filter_products(data_products,
+                    productType=["AUXILIARY",],
+                    extension="fits",
+                    productSubGroupDescription = "GS-FG",
+                    dataproduct_type='image'
+        )
+
         if obs_num != 'None' and visit != 'None':
-            matched_obs = Observations.query_criteria(
-                obs_collection = 'JWST',
-                proposal_id=str(pid),
-            )
-            
-            data_products = [Observations.get_product_list(obs) for obs in matched_obs]
-            data_products = vstack(data_products)
 
-            products = Observations.filter_products(data_products,
-                        productType=["AUXILIARY",],
-                        extension="fits",
-                        productSubGroupDescription = "GS-FG",
-                        dataproduct_type='image'
-            )
-
-            
             obs_num_col = []
             visit_col = []
 
@@ -418,23 +426,8 @@ class load:
 
             mask3 = productsx['visit'] == int(visit)
             productsx = productsx[mask3]
-            
 
         elif obs_num != 'None':
-            matched_obs = Observations.query_criteria(
-                obs_collection = 'JWST',
-                proposal_id=str(pid),
-            )
-
-            data_products = [Observations.get_product_list(obs) for obs in matched_obs]
-            data_products = vstack(data_products)
-
-            products = Observations.filter_products(data_products,
-                        productType=["AUXILIARY",],
-                        extension="fits",
-                        productSubGroupDescription = "GS-FG",
-                        dataproduct_type='image'
-            )
 
             obs_num_col = []
 
@@ -449,22 +442,7 @@ class load:
             mask2 = productsx['obs_num'] == int(obs_num)
             productsx = productsx[mask2]
 
-
         else:
-            matched_obs = Observations.query_criteria(
-                obs_collection = 'JWST',
-                proposal_id=str(pid),
-                )
-
-            data_products = [Observations.get_product_list(obs) for obs in matched_obs]
-            data_products = vstack(data_products)
-
-            products = Observations.filter_products(data_products,
-                                    productType=["AUXILIARY",],
-                                    extension="fits",
-                                    productSubGroupDescription = "GS-FG",
-                                    dataproduct_type='image'
-                                    )
         
[…]
