### 2024-04-09 GalagaBits/JWST-FGS-Spelunker#37 @GalagaBits
@evamariaa Should I provide a brief description of each main features? Or make different sections for each main feature?
