### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#35 @GalagaBits
@taylorbell57 For the forced plotting of mnemonics, mnemonics_local, and periodogram, in a future update I will separate the plotting element into a separate plotting functions mnemonics_plot, mnemonics_local_plot, and periodogram_plot in case the user wants to generate a plot without rerunning the periodogram and mnemonics functions. I plan to rework the mnemonics functions to easily allow any mnemonics in a single attribute, but this will take some time for me to do at the moment
