### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#35 @GalagaBits
@taylorbell57 I have recently added the feature to control the number of bin parameters.
