### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#31 @GalagaBits
@taylorbell57 I will remove `guidestar_targets` because this was an older notebook and its main point, which is to advertise the guidestar_plot feature, is already shown in the quickstart guide
