### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#31 @GalagaBits
@taylorbell57 I have also fixed alot of the images with the new change in website layout, but I still need to tweak some more. For the most part everything in the docs now should be readable
