### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#31 @GalagaBits
I will keep the pixel_centroid_mnemonics.rst webpage since someone might not notice the notebook if they are using the readthedocs instead of the github
