### 2024-04-16 be06738
added binned features
