+++ docs/guidestar_targets.rst
@@ -1,255 +1,0 @@
-Guidestar Targets Example
-=========================
-
-Some programs uses multiple guidestar targets for a given observation.
-We can observe the technical functions of the JWST when FGS locks on to
-a guidestar and switches to the next using Spelunker. For this example,
-we will be exploring Program ID 6079.
-
-.. code:: ipython3
-
-    import sys
-    sys.path.append('/Users/ddeal/JWST-FGS-Spelunker/JWST-FGS-spk-main/src/')
-    
-    import spelunker
-    
-    spk = spelunker.load('/Users/ddeal/JWST-Treasure-Chest/', pid=1534)
-
-
-
-
-.. code:: ipython3
-
-    spk.guidestar_plot()
-
-
-
-
-.. parsed-literal::
-
-    <WCSAxes: title={'center': 'Guidestar positions — 1534'}>
-
-
-
-
-.. image:: guidestar_targets_files/guidestar_targets_3_1.png
-   :scale: 60%
-
-.. code:: ipython3
-
-    spk.object_properties
-
-
-
-
-.. raw:: html
-
-    <div>
-    <style scoped>
-        .dataframe tbody tr th:only-of-type {
-            vertical-align: middle;
-        }
-    
-        .dataframe tbody tr th {
-            vertical-align: top;
-        }
-    
-        .dataframe thead th {
-            text-align: right;
-        }
-    </style>
-    <table border="1" class="dataframe">
-      <thead>
-        <tr style="text-align: right;">
-          <th></th>
-          <th>guidestar_catalog_id</th>
-          <th>gaiadr1ID</th>
-          <th>gaiadr1ID</th>
-          <th>int_start</th>
-          <th>int_stop</th>
-          <th>ra</th>
-          <th>dec</th>
-          <th>Jmag</th>
-          <th>Hmag</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <th>0</th>
-          <td>S1HP079555</td>
-          <td>4658077781377287680</td>
-          <td>4658077781376437888</td>
-          <td>59917.066396</td>
-          <td>59917.074354</td>
-          <td>80.837584</td>
-          <td>-69.541124</td>
-          <td>13.659</td>
-          <td>12.898</td>
-        </tr>
-        <tr>
-          <th>1</th>
-          <td>S1HP080554</td>
-          <td>4658077991763987712</td>
-          <td>4658077991799023616</td>
-          <td>59917.089163</td>
-          <td>59917.096759</td>
-          <td>80.806837</td>
-          <td>-69.530972</td>
-          <td>15.001</td>
-          <td>14.282</td>
-        </tr>
-        <tr>
-          <th>2</th>
-          <td>S1HP078573</td>
-          <td>4657983910572904360</td>
-          <td>4657983910572904360</td>
-          <td>59917.112547</td>
-          <td>59917.118705</td>
-          <td>80.807043</td>
-          <td>-69.553474</td>
-          <td>13.839</td>
-          <td>13.078</td>
-        </tr>
-        <tr>
-          <th>3</th>
-          <td>S1HP079590</td>
-          <td>4657986831103727872</td>
-          <td>4657986835382986016</td>
-          <td>59918.999015</td>
-          <td>59919.005848</td>
-          <td>80.510790</td>
-          <td>-69.545479</td>
-          <td>15.410</td>
-          <td>14.839</td>
-        </tr>
-        <tr>
-          <th>4</th>
-          <td>S1HP079769</td>
-          <td>4657986831078160832</td>
-          <td>4657986835433225728</td>
[…]
