### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#35 @GalagaBits
Actually, I have incorporated the ability to turn on and off plots in `mnemonics`, `mnemonics_local`, and `periodogram`.
