+++ src/spelunker.py
@@ -47,78 +47,10 @@ class load:
 
                 >>> spk = spelunker.load(pid=1534,obs_num='2',token='abcdefg1234567')
 
-    :param dir: (optional, string)
-        The parameter "dir" allows the user to change the directory of the script. Changing the directory dictates where the class loads, creates folders, and save data and results.
-        If ignored, "load" will save the current directory as an attribute for use with other functions.
-
-    :param pid: (optional, int or string)
-        An Python ``int`` or ``string`` of the Program ID for a given JWST program (i.e.: 1296, '1296'). Program IDs that are publicly avaliable are able to be accessed. Program IDs 
-        with exclusive rights will not be accessible unless the user has a MAST API token containing the rights to the program ID. The MAST API token must be entered with param "token"
-        to use programs marked with exclusive rights. The program ID needs to be entered to use all of functions in "load". However, the pid can be ignored to use specific functions
-        within the class, though functionality will be severely limited. You can also forgo inputting the program ID to create a 'blank' object, then later input data for a pid using
-        the function ``download`` given ``pid`` and the following parameters. For example:
+    Parameters
+    ----------
                 
-                >>> spk = spelunker.load()
-                >>> spk.download(pid=1534)
-
-    :param obs_num: (optional, int or string)
-        The integer or string of the observation number of a program ID. Inputting the observation number will filter the auxiliary FITS to only include the files with the given
-        number before downloading all files. Note that when you specifify the observation number, you must also input the ``pid``. 
-        
-    :param visit: (optional, int or string)
-        The visit parameter requires a Python ``int`` or ``string``. Inputting the visit number will fillter the auxiliary FITS to only include the files with the given parameter
-        before downloading all files.
-
-    :param visit_group: (optional, int or string)
-        The visit group parameter requires either a ``int`` or ``string``. After the FITS files downloads for a corresponding program ID, specifying ``vist_group`` will filter the
-        data to only include files with the given visit group before loading the data into the object.
-
-    :param parallel_sequence_id: (optional, int or string)
-        The parallel sequence ID parameter requires either a ``int`` or ``string``. After the FITS files downloads for a corresponding program ID, specifying ``parallel_sequence_id`` 
-        will filter the data to only include files with the given parallel sequence ID before loading the data into the object.
-
-    :param activity_number: (optional, int or string)
-        The activity number requires either a ``int`` or ``string``. After the FITS files downloads for a corresponding program ID, specifying ``activity_number`` will filter the
-        data to only include files with the given activity number before loading the data into the object.
-
-    :param exposure_number: (optional, int or string)
-        The exposure_number requires either a ``int`` or ``string``. After the FITS files downloads for a corresponding program ID, specifying ``exposure_number`` will filter the
-        data to only include files with the given exposure number before loading the data into the object.
-
-    :param dir_seg: (optional, int or string)
-        The segment number requires a Python ``int`` or ``string``. Specifying the segment number will include guidestar FITS with only the input number in the header if the segmented
-        files has 'seg' within the naming scheme for the filename. This parameter will filter associated program ID FITS after downloading has completed and before the data is loaded
-        into the object.
-
-    :param guider: (optional, int or string)
-        Similarly to `dir_seg`, the guider number requires a Python ``int`` or ``string``. Specifying the guider number will include guidestar FITS with only the input number in the 
-        header if the files has 'guider' within the naming scheme for the filename. This parameter will filter associated program ID FITS after downloading has completed and before 
-        the data is loaded into the object.
-
-    :param calib_level: (optional, int or string)
-        The calibration level can be specified as a Python ``int`` or ``string``. This parameter determines the calibration level for the FG-GS FITS. The only inputs for this
[…]
