### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#26 @GalagaBits
@evamariaa Yes! You can still use Spelunker without using any astroquery related functions from Spelunker. Once you have downloaded the FITS for the GS-FG guidestar data, you can use the `readfile` function to load in all files with a specified PID in the parameter of the function. You will get all the features of Spelunker without the astroquery Spelunker functions. The latest Spelunker installation v1.1.10 works the best and avoids some of the bugs I recently went through. You'll need to change into the directory with your FITS files and use

```
import spelunker

spk = spelunker.load()

spk.readfile(pid=4098)

```

to get all of your guidestar data loaded into spelunker. You can also load data with selected visit and observation numbers with the respective parameters.

As for the astroquery functions not working properly, I don't know why it is not working properly yet, but I do know that Spelunker get a list of data products from MAST given a query criteria:

```
        matched_obs = Observations.query_criteria(
            obs_collection = 'JWST',
            proposal_id=str(pid),
        )

```.

All of the data products links is then read according to `matched_obs` if it is from JWST and if it is under your program id. I realize that this takes all of the data products and can be very slow. I have tried to make it faster by looking for parameters that could successfully query a smaller amount of data products. The data products list is read and filtered using the following snippet:

```
        data_products = [Observations.get_product_list(obs) for obs in matched_obs]
        data_products = vstack(data_products)

        products = Observations.filter_products(data_products,
                    productType=["AUXILIARY",],
                    extension="fits",
                    productSubGroupDescription = "GS-FG",
                    dataproduct_type='image'
        )

 ```
It is further filtered with astropy table masking according to your input parameters. After the table of data products links has been filtered to a new table (`productsx`), the actual FITS data is downloaded:

```
manifest = Observations.download_products(productsx, download_dir=self.directory)
```

Let me know if you have any comments to this by pinging me and I will answer to the best of my ability
