### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#29 @nespinoza
Hi @taylorbell57! Thanks a lot for the feedback --- I implemented most of it, except for:

- We didn't implement your comment about the custom function `bin_fgs_to_science`, as there is no such function in the `spelunker` codebase itself. We _could_ add it, but the function is a bit ad-hoc right now --- we'll think of a better way to implement this in the future to the actual `spelunker` codebase, and update the notebook accordingly.
- Similarly, we have explained the messages that `ray` spits out, but we have not implemented any `verbose` key into the function for now. We will in the future.
- For `bin_data`, we could --- but it might obscure what the function actually does in the background (in parallel with what happens with `bin_fgs_to_science`) --- so better to keep as is.

In addition, for your second to last point, added some wording on why Y is better in this case: "(...) Interestingly, in this case, the Y-standard deviation samples the event even better, when considering the amplitude of the 'jump' in this time series to the overall scatter of the time-series for the Y-standard deviation itself:".

Thanks again! Will mark as done.

N.
