### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#31 @GalagaBits
Actually, I'll remove the pixel_centroid_mnemonics.rst for now
