### 2024-04-16 8acc381
new changes to pixel_centroid
