### 2024-04-16 GalagaBits/JWST-FGS-Spelunker#31 @GalagaBits
Okay, I have addressed each issue
