### 2024-04-17 85ba015
Added image to paper
