+++ paper/paper.bib
@@ -1,45 +1,67 @@
 
 @article{gardner_james_2023,
-	title = {The James Webb Space Telescope Mission},
+	title = {The {James} {Webb} {Space} {Telescope} {Mission}},
 	volume = {135},
 	issn = {0004-6280},
 	url = {https://ui.adsabs.harvard.edu/abs/2023PASP..135f8001G},
 	doi = {10.1088/1538-3873/acd1b5},
-	abstract = {Twenty-six years ago a small committee report, building on earlier studies, expounded a compelling and poetic vision for the future of astronomy, calling for an infrared-optimized space telescope with an aperture of at least 4 m. With the support of their governments in the {US}, Europe, and Canada, 20,000 people realized that vision as the 6.5 m James Webb Space Telescope. A generation of astronomers will celebrate their accomplishments for the life of the mission, potentially as long as 20 yr, and beyond. This report and the scientific discoveries that follow are extended thank-you notes to the 20,000 team members. The telescope is working perfectly, with much better image quality than expected. In this and accompanying papers, we give a brief history, describe the observatory, outline its objectives and current observing program, and discuss the inventions and people who made it possible. We cite detailed reports on the design and the measured performance on orbit.},
-	pages = {068001},
-	journaltitle = {Publications of the Astronomical Society of the Pacific},
-	author = {Gardner, Jonathan P. and Mather, John C. and Abbott, Randy and Abell, James S. and Abernathy, Mark and Abney, Faith E. and Abraham, John G. and Abraham, Roberto and Abul-Huda, Yasin M. and Acton, Scott and Adams, Cynthia K. and Adams, Evan and Adler, David S. and Adriaensen, Maarten and Aguilar, Jonathan Albert and Ahmed, Mansoor and Ahmed, Nasif S. and Ahmed, Tanjira and Albat, Rüdeger and Albert, Loïc and Alberts, Stacey and Aldridge, David and Allen, Mary Marsha and Allen, Shaune S. and Altenburg, Martin and Altunc, Serhat and Alvarez, Jose Lorenzo and Álvarez-Márquez, Javier and Alves de Oliveira, Catarina and Ambrose, Leslie L. and Anandakrishnan, Satya M. and Andersen, Gregory C. and Anderson, Harry James and Anderson, Jay and Anderson, Kristen and Anderson, Sara M. and Aprea, Julio and Archer, Benita J. and Arenberg, Jonathan W. and Argyriou, Ioannis and Arribas, Santiago and Artigau, Étienne and Arvai, Amanda Rose and Atcheson, Paul and Atkinson, Charles B. and Averbukh, Jesse and Aymergen, Cagatay and Bacinski, John J. and Baggett, Wayne E. and Bagnasco, Giorgio and Baker, Lynn L. and Balzano, Vicki Ann and Banks, Kimberly A. and Baran, David A. and Barker, Elizabeth A. and Barrett, Larry K. and Barringer, Bruce O. and Barto, Allison and Bast, William and Baudoz, Pierre and Baum, Stefi and Beatty, Thomas G. and Beaulieu, Mathilde and Bechtold, Kathryn and Beck, Tracy and Beddard, Megan M. and Beichman, Charles and Bellagama, Larry and Bely, Pierre and Berger, Timothy
[…]
