### 2024-04-17 119d482
Added image to paper
