+++ notebooks/examples/pixel_centroid_mnemonics.rst
@@ -1,289 +1,0 @@
-Pixel centroid changes and mnemonics \| PID 2079 Example
-========================================================
-
-.. code:: ipython3
-
-    import sys
-    sys.path.append('/Users/ddeal/JWST-FGS-Spelunker/JWST-FGS-Spelunker-main/src/')
-    import spelunker
-    
-    spk = spelunker.load('/Users/ddeal/JWST-Treasure-Chest/', pid=2079, obs_num=4, visit=1)
-
-
-.. parsed-literal::
-
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031214828-seg001_cal.fits with expected size 51960960. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031214828-seg002_cal.fits with expected size 43096320. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031214829-seg001_cal.fits with expected size 64540800. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031214829-seg002_cal.fits with expected size 53496000. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031231651-seg003_cal.fits with expected size 51960960. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031231651-seg004_cal.fits with expected size 31645440. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031231652-seg003_cal.fits with expected size 64540800. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023031231652-seg004_cal.fits with expected size 39245760. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023032005233-seg005_cal.fits with expected size 51960960. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023032005233-seg006_cal.fits with expected size 38445120. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023032005234-seg005_cal.fits with expected size 64540800. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023032005234-seg006_cal.fits with expected size 47736000. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023032005730_cal.fits with expected size 5129280. [astroquery.query]
-    INFO: Found cached file ./mastDownload/JWST/jw02079004001_02101_00001_nis/jw02079004001_gs-fg_2023032010230_cal.fits with expected size 5137920. [astroquery.query]
[…]
