### 2024-04-17 fb4e1a4
Added image to paper
