+++ paper/paper.md
@@ -32,7 +32,7 @@ In addition, ``spelunker`` provides visualization and analysis tools to study th
 
 The James Webb Space Telescope [@gardner_james_2023] produces some of the highest sensitivity imaging of the cosmos across all instruments. One of them, the NIRISS Fine Guidance Sensor [@doyon_jwst_2012], provides guide star imaging with a passband of 0.6 to 5 microns through two separate channels, each with a $2.3’ \times 2.3’$ field of view (FOV) and a sampling rate of 64 ms—data that is taken in parallel and is thus available for every JWST observing program. While the onboard system uses guidestars to guide the attitude control system (ACS) which stabilizes the observatory, the astronomical community can also use the data products associated with these 64 ms cadence images as science products. Usages range from studying guide star photometry in search of transient phenomena to using this data to identify and investigate technical anomalies that might occur during scientific observations with the rest of the JWST instruments. Despite this wide range of possible usages, these data products are not straightforward to manipulate and analyze, and there is no publicly available package to download, investigate, and research guidestar data. ``spelunker`` is a Python library that was developed to enable access to these guide star data products and their analysis.
 
-![A snippet from the guidestar timeseries from Cycle 1 GO Program ID 1803. **Top** — The guidestar timeseries of PID 1803 after loading it into ``spelunker`` using ``timeseries_binned_plot``. **Bottom** — The same timeseries after applying pixel level decorrelation (PLD) using ``optimize_photometry`` [@deming_spitzer_2015]. \label{fig:guidestar_1803}](pid1803-001-001_optimize.png)
+![A snippet from the guidestar timeseries from Cycle 1 GO Program ID 1803, observation 1, and visit 1. **Top** — The guidestar timeseries of PID 1803 after loading it into ``spelunker`` using ``timeseries_binned_plot``. The timeseries uses the sum of counts in each guidestar fine guidence (GS-FG) frame. The data has no significant features. **Middle** — The same timeseries after applying pixel level decorrelation (PLD) using ``optimize_photometry`` [@deming_spitzer_2015]. There are now prominent decreases in flux which were previously unseen with the raw timeseries data. **Bottom** — Gaussian fitted x pixel coordinate and y pixel coordinate for each frame in this section of timeseries data. The guidestar shifts around in this timeseries, likely highlighting the functions of the ACS. \label{fig:guidestar_1803}](timeseries_plot.png)
 
 ![There are seven parameters `gauss2d_fit` measures: amplitude (counts of the guidestar), x pixel coordinate, y pixel coordinate, the x and y standard deviations, theta (orientation of the Gaussian model), and the offset (the background counts). This diagram visualizes what each parameter represents on the Gaussian model. \label{fig:Gaussian_diagram}](Gaussian_diagram.png)
 
