### 2024-04-17 adba041
Added image to paper
