### 2024-04-17 bfa8457
Added image to paper
