### 2024-04-17 c425f75
Added image to paper
