### 2024-04-17 596d039
Added image to paper
