### 2024-04-17 1557a23
Added image to paper
