### 2024-04-17 50ebd8f
Added image to paper
