### 2024-04-18 ab1c0e3
Added image to paper
