+++ paper/paper.bib
@@ -65,3 +65,41 @@ ADS Bibcode: 2015ApJ...805..132D},
 	pages = {132},
 	file = {Full Text PDF:/Users/galagabits/Zotero/storage/WFVIW7DX/Deming et al. - 2015 - Spitzer Secondary Eclipses of the Dense, Modestly-irradiated, Giant Exoplanet HAT-P-20b Using Pixel-.pdf:application/pdf},
 }
+
+@article{scargle_studies_1982,
+	title = {Studies in astronomical time series analysis. {II}. {Statistical} aspects of spectral analysis of unevenly spaced data.},
+	volume = {263},
+	issn = {0004-637X},
+	url = {https://ui.adsabs.harvard.edu/abs/1982ApJ...263..835S},
+	doi = {10.1086/160554},
+	abstract = {Detection of a periodic signal hidden in noise is frequently a goal in astronomical data analysis. This paper does not introduce a new detection technique, but instead studies the reliability and efficiency of detection with the most commonly used technique, the periodogram, in the case where the observation times are unevenly spaced. This choice was made because, of the methods in current use, it appears to have the simplest statistical behavior. A modification of the classical definition of the periodogram is necessary in order to retain the simple statistical behavior of the evenly spaced case. With this modification, periodogram analysis and least-squares fitting of sine waves to the data are exactly equivalent. Certain difficulties with the use of the periodogram are less important than commonly believed in the case of detection of strictly periodic signals. In addition, the standard method for mitigating these difficulties (tapering) can be used just as well if the sampling is uneven. An analysis of the statistical significance of signal detections is presented, with examples},
+	urldate = {2024-04-18},
+	journal = {The Astrophysical Journal},
+	author = {Scargle, J. D.},
+	month = dec,
+	year = {1982},
+	note = {Publisher: IOP
+ADS Bibcode: 1982ApJ...263..835S},
+	keywords = {Astronomy, Fourier Transformation, Frequency Response, Power Spectra, Signal Detection, Signal To Noise Ratios, Spectrum Analysis, Statistical Distributions, Time Series Analysis},
+	pages = {835--853},
+	file = {Full Text PDF:/Users/galagabits/Zotero/storage/765TL2HX/Scargle - 1982 - Studies in astronomical time series analysis. II. Statistical aspects of spectral analysis of uneven.pdf:application/pdf},
+}
+
+@article{lomb_least-squares_1976,
+	title = {Least-{Squares} {Frequency} {Analysis} of {Unequally} {Spaced} {Data}},
+	volume = {39},
+	issn = {0004-640X},
+	url = {https://ui.adsabs.harvard.edu/abs/1976Ap&SS..39..447L},
+	doi = {10.1007/BF00648343},
+	abstract = {The statistical properties of least-squares frequency analysis of unequally spaced data are examined. It is shown that, in the least-squares spectrum of gaussian noise, the reduction in the sum of squares at a particular frequency is aX \{2/2\} variable. The reductions at different frequencies are not independent, as there is a correlation between the height of the spectrum at any two frequencies,f 1 andf 2, which is equal to the mean height of the spectrum due to a sinusoidal signal of frequencyf 1, at the frequencyf 2. These correlations reduce the distortion in the spectrum of a signal affected by noise. Some numerical illustrations of the properties of least-squares frequency spectra are also given.},
+	urldate = {2024-04-18},
+	journal = {Astrophysics and Space Science},
+	author = {Lomb, N. R.},
+	month = feb,
+	year = {1976},
+	note = {Publisher: Springer
+ADS Bibcode: 1976Ap\&SS..39..447L},
[…]
