### 2024-04-18 GalagaBits/JWST-FGS-Spelunker#37 @GalagaBits
Okay, I have added your feed back in the paper!
