### 2024-04-18 #6202 @GalagaBits
@taylorbell57 The current version is v1.1.11. This should now be reflected in GitHub and PyPI
