+++ paper/paper.bib
@@ -47,25 +47,6 @@ ADS Bibcode: 2018SPIE10704E..13M},
 	pages = {1070413},
 }
 
-@article{deming_spitzer_2015,
-	title = {Spitzer {Secondary} {Eclipses} of the {Dense}, {Modestly}-irradiated, {Giant} {Exoplanet} {HAT}-{P}-20b {Using} {Pixel}-level {Decorrelation}},
-	volume = {805},
-	issn = {0004-637X},
-	url = {https://ui.adsabs.harvard.edu/abs/2015ApJ...805..132D},
-	doi = {10.1088/0004-637X/805/2/132},
-	abstract = {HAT-P-20b is a giant metal-rich exoplanet orbiting a metal-rich star. We analyze two secondary eclipses of the planet in each of the 3.6 and 4.5 μm bands of Warm Spitzer. We have developed a simple, powerful, and radically different method to correct the intra-pixel effect for Warm Spitzer data, which we call pixel-level decorrelation (PLD). PLD corrects the intra-pixel effect very effectively, but without explicitly using—or even measuring—the fluctuations in the apparent position of the stellar image. We illustrate and validate PLD using synthetic and real data and comparing the results to previous analyses. PLD can significantly reduce or eliminate red noise in Spitzer secondary eclipse photometry, even for eclipses that have proven to be intractable using other methods. Our successful PLD analysis of four HAT-P-20b eclipses shows a best-fit blackbody temperature of 1134 ± 29 K, indicating inefficient longitudinal transfer of heat, but lacking evidence for strong molecular absorption. We find sufficient evidence for variability in the 4.5 μm band that the eclipses should be monitored at that wavelength by Spitzer, and this planet should be a high priority for James Webb Space Telescope spectroscopy. All four eclipses occur about 35 minutes after orbital phase 0.5, indicating a slightly eccentric orbit. A joint fit of the eclipse and transit times with extant RV data yields ecos ω =0.01352-0.00057+0.00054 and establishes the small eccentricity of the orbit to high statistical confidence. HAT-P-20b is another excellent candidate for orbital evolution via Kozai migration or other three-body mechanisms.},
-	urldate = {2024-04-17},
-	journal = {The Astrophysical Journal},
-	author = {Deming, Drake and Knutson, Heather and Kammer, Joshua and Fulton, Benjamin J. and Ingalls, James and Carey, Sean and Burrows, Adam and Fortney, Jonathan J. and Todorov, Kamen and Agol, Eric and Cowan, Nicolas and Desert, Jean-Michel and Fraine, Jonathan and Langton, Jonathan and Morley, Caroline and Showman, Adam P.},
-	month = jun,
-	year = {2015},
-	note = {Publisher: IOP
-ADS Bibcode: 2015ApJ...805..132D},
-	keywords = {Astrophysics - Earth and Planetary Astrophysics, eclipses, infrared: planetary systems, planetary systems, planets and satellites: atmospheres},
-	pages = {132},
-	file = {Full Text PDF:/Users/galagabits/Zotero/storage/WFVIW7DX/Deming et al. - 2015 - Spitzer Secondary Eclipses of the Dense, Modestly-irradiated, Giant Exoplanet HAT-P-20b Using Pixel-.pdf:application/pdf},
-}
-
 @article{scargle_studies_1982,
 	title = {Studies in astronomical time series analysis. {II}. {Statistical} aspects of spectral analysis of unevenly spaced data.},
 	volume = {263},
