### 2024-04-19 #6202 @taylorbell57
@GalagaBits, how easy would it be to add a progress bar to the `gauss2d_fit` function? I'm testing the `fgs-spelunker-and-tsos.ipynb` notebook out right now, and I can't tell if the function call has crashed or it is just taking a long time since there is no progress bar
