### 2024-04-19 #6202 @GalagaBits
@taylorbell57, I'll have to research on how to implement a progress, so that will take a while since I am currently in finals season. It's likely taking a while due to the large amount of frames.
