### 2024-04-19 #6202 @taylorbell57
So it seemed my first attempt had indeed crashed (took >45 minutes), but in trying to understand what happened I found a way to get a progress bar; I'll submit a pull request to enable that useful feature. The main issue though is that you're currently using ray to parallelize each individual fit rather than running multiple fits in parallel. Each individual fit takes far less than a second, so this method of parallelization offers little-to-no gains (in fact, using a tool like `htop` I can see that regardless of what I set `ncpus` to, the function only really uses ~1 thread since it finishes too fast to use any more threads). Now that I can tell that the function is actually working (with my temporarily hacked-in progress bar), neither of these issues are something that should hold back the publication of the paper. I'll just open issues and/or pull requests for them on your GitHub.

I'll continue reviewing/testing the other components of your package, paper, and documentation - looking good so far!
