### 2024-04-19 #6202 @GalagaBits
Thank you for that! Also, I looked at Activity Monitor when ever Spelunker runs Gaussian fits, the usage does not seem very high. Using more cores with `ray` somewhat speed up the process a bit from my experience, but I feel that I can implement ray better. Splitting the data into chunks depending on the available cores/threads and running each chunk with ray might be more efficient and much faster. Do you have any suggestions?

And feel free to tell me about any questions or comments!
