### 2024-04-19 #6202 @taylorbell57
@GalagaBits soon I'll open an issue on the Spelunker repo to discuss this in detail (since it's not really relevant to this review anymore) and provide tips on how you could (likely) greatly speed this function up, but for reference the loop at line [spelunker.py#L1260](https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/fc769668ea39efe4c759afafe7d4b680e48f681b/src/spelunker.py#L1260) took me 8m4s with the full 28 threads available on my CPU (a desktop Intel i7-14700K) and 7m51s with only 1 thread, while the loop at [spelunker.py#L1272](https://github.com/GalagaBits/JWST-FGS-Spelunker/blob/fc769668ea39efe4c759afafe7d4b680e48f681b/src/spelunker.py#L1272) took 18m46s with 28 threads and 19m04s with only 1 thread
