### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
@GalagaBits, I see that the jwstuser package is still not included in the setup.py file. Had you just forgotten about this GitHub issue or did my suggestion not work for you?
