### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#25 @GalagaBits
@taylorbell57, I tried implementing all suggestions here, but I ran into the same error.
