### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#43 @taylorbell57
Increased accessibility of the readthedocs link from GitHub
On the main JWST-FGS-Spelunker GitHub page, on the right panel beside **About** you should see a gear - if you click on that, then there should be a `Website` field where you can enter `https://jwst-fgs-spelunker.readthedocs.io` which will make it easier for future users to find the link to your readthedocs page.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
