### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#43 @GalagaBits
Ah I was wondering where to find that in the previous issue, so I put the link in the first lines of the readme. I see it now and I have added it to the About section
