### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
Ah indeed, I just tried it myself and got the same error. After some searching around, I found an explanation in [PEP-0440](https://peps.python.org/pep-0440/#direct-references).

> Public index servers SHOULD NOT allow the use of direct references in uploaded distributions. Direct references are intended as a tool for software integrators rather than publishers.
