### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#25 @taylorbell57
I guess you'll have to wait until `jwstuser` is uploaded to PyPI before you could include it. Thanks for letting me know - this is relevant to some of my own repos which I've yet to try to host on PyPI
