### 2024-04-19 GalagaBits/JWST-FGS-Spelunker#44 @taylorbell57
Unused arguments to readfile function
There are numerous currently unused arguments to the `readfile` function, specifically all of
```
visit_group='None', parallel_sequence_id='None',  activity_number='None', exposure_number='None', dir_seg='None', guider='None'
```
are not currently used in the function. It seems these arguments were previously used but the relevant code has currently been commented out. Most/all of these arguments are used in the `download` function, so is there any reason they're no longer used in the `readfile` function while the arguments are still available? Unless there is a clear reason to keep things as they currently are, I recommend you either enable the use of these arguments or remove them from the `readfile` arguments list.

This issue was made as part of the review for https://github.com/openjournals/joss-reviews/issues/6202
