### 2024-04-19 #6202 @taylorbell57
Alright, things are looking much improved. Outside of the newly opened GitHub issues linked above this message, the only things that I think still need to be addressed are:
- [ ] You should ensure that you've deleted all `.DS_Store` files, including `notebooks/.DS_Store`, `plots/.DS_Store`
- [ ] You should still add a bit of text to your paper to address our checklist item which reads "State of the field: Do the authors describe how this software compares to other commonly-used packages?". You now state that you make use of astroquery and MAST, but a sentence or two about "how this software compares to other commonly-used packages" would still be useful, for example by explaining how the tool differs from (but may be useful to others who are using) packages like STScI's `jwst` package or other open-source JWST data reduction and/or fitting packages.

After that's all taken care of, then I think you should be good to go!
