### 2024-04-19 3b45269
Adding tqdm progress bars to slow parts of gauss2d_fit
This partially takes care of #42 by at least giving the progress status (estimated runtime, percentage complete) for the two slow loops of the `gauss2d_fit` function using `tqdm`'s progress bar utility, but fixing the currently non-functional parallelization of the function still remains to be done
