### 2024-04-20 GalagaBits/JWST-FGS-Spelunker#41 @GalagaBits
@taylorbell57 For some reason, the mnemonics are not lining up as they did before. I'll have to look into this, but this will take me a while
