+++ docs/pixel_centroid_mnemonics.rst
@@ -5,18 +5,26 @@ For this example, we will use the JWST engineering telemetry database to overplo
 Mnemonics are useful in detecting anomalies in guidestar data and differentiating technical events on the JWST from the guidestar flux. 
 We will use PID 2079, observation number 4, and visit number 1 and find out if there are any technical anomalies in our data using Spelunker.
 
+First, load the program ID. The download for this PID is long, we are using `spk.readfile` to load in the files that are already in the directory.
+
+
 .. code:: ipython3
 
-    import spelunker
-    
-    spk = spelunker.load('/Users/ddeal/JWST-Treasure-Chest/', pid=2079, obs_num=4, visit=1)
+import spelunker
 
+    spk = spelunker.load()
+    spk.readfile(pid=2079, obs_num=4, visit=1)
+
+    '''
+    Use spelunker.load here instead of `spk.readfile` to download the files.
+    '''
+    #spk = spelunker.load(pid=2079, obs_num=4, visit=1)
 
 You will need to input your MAST API Token to access JWST Engineering Telemetry and mnemonics.
 
 .. code:: ipython3
 
-    spk.mast_api_token = 'enter_mast_token_id_here'
+    spk.mast_api_token = 'MAST_API_TOKEN'
 
 Lets look at how many guidestars this program uses.
 
@@ -91,17 +99,15 @@ flux timeseries for the guidestar in the program.
 
 
 .. image:: pixel_centroid_mnemonics_files/pixel_centroid_mnemonics_6_0.png
-   :scale: 40%
+   :width: 800pt
 
-We can see that the timeseries is broken up into sepeate pieces. Lets see how the obserbed properties of the target
-changes overtime. We fitted gaussians to each frame using
-``spk.gauss2d_fit`` and saved the outpput as an astropy table ``dat``
-file, so we can load the table here.
+We can see that the timeseries is broken up into sepeate pieces. Lets see how the pixel cooridinates, pixel standard deviation and Gaussian fitted amplitude of the guidestar changes overtime. We fitted gaussians to each frame using `spk.gauss2d_fit` and saved the outpput as an astropy table `dat` file, so we can load the table here.
 
 .. code:: ipython3
 
     from astropy.io import ascii
-    gaussfit_table = ascii.read('/Users/ddeal/Spelunker-older/JWST-Treasure-Chest-2023/pid2079_observation04_visit1.dat')
+    gaussfit_table = ascii.read('/Users/galagabits/Developer/FGS-Spelunker/JWST-FGS-Spelunker/notebooks/examples/data/pid2079_observation04_visit1_short.dat') # Navigate to the included Gaussian results file
+    time = spk.fg_time[1300000:1900000] # The Gaussian table was cut short to fit in GitHub, so the time needs to be the same shape as well.
 
 .. code:: ipython3
 
@@ -110,7 +116,7 @@ file, so we can load the table here.
 
 
 .. image:: pixel_centroid_mnemonics_files/pixel_centroid_mnemonics_9_0.png
-   :scale: 40%
+   :width: 800pt
 
 
 
@@ -127,108 +133,110 @@ file, so we can load the table here.
 
 
 We can clearly see that there are some periodic changes in the centroids
-every 0.2mjd with even smaller changes within. Lets zoom into one of the
+every hour with even smaller changes within. Lets zoom into one of the
 centroid plots.
 
[…]
