### 2024-04-21 GalagaBits/JWST-FGS-Spelunker#44 @GalagaBits
Okay, this has been fixed in the latest update
