### 2024-04-21 GalagaBits/JWST-FGS-Spelunker#41 @GalagaBits
@taylorbell57 I have fixed the reproducibility of this example. Also, I removed the animation section for now to keep the focus on mnemonics.
