+++ paper/paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 ``spelunker`` is a Python library that provides several tools for analyzing and visualizing JWST NIRISS FGS guide star data products. The library can download guide star data from the Mikulski Archive for Space Telescopes [@marston_overview_2018] for any given Program ID in a single line of code. Through efficient parallelization processes, the pipeline can use this data to extract photometry in seconds and point-spread function information from all frames within a selected Program ID in minutes. Data loaded within the package allows for many possible analyses such as Gaussian fitting, periodograms of each fitted Gaussian parameter and flux timeseries, and photometry optimization.
 
-In addition, ``spelunker`` provides visualization and analysis tools to study these products in detail, including the incorporation of JWST engineering telemetry, which can be used to put the JWST FGS data products in context with other observatory variables that might help explain data patterns both in the primary science data products and on the JWST FGS data itself. The library also cross-references, tracks and stores guide star metadata for the user. This metadata includes information such as the GAIA ID of the guide star, coordinates, and magnitudes, among others. This pipeline empowers users to study guidestar data with applications like technical anomaly detection and time-domain astronomy at a high 64 ms cadence.
+In addition, ``spelunker`` provides visualization and analysis tools to study these products in detail, including the incorporation of JWST engineering telemetry, which can be used to put the JWST FGS data products in context with other observatory variables that might help explain data patterns both in the primary science data products and on the JWST FGS data itself. The library also cross-references, tracks and stores guide star metadata for the user. This metadata includes information such as the GAIA ID of the guide star, coordinates, and magnitudes, among others. This pipeline empowers users to study guidestar data with applications like technical anomaly detection and time-domain astronomy at a high 64 ms cadence. Existing pipeline such as ``jwst`` provides the structure for calibrataing NIRISS/FGS, NIRSpec, NIRCam, and MIRI data, such as with JWST ``Datamodels``. ``spelunker``  tools are engineered to explore FGS guidestars in two main categories: as a technique to detect technical anomalies in JWST data, and as a pipeline to study time-domain astronomy with guidestars.
 
 # Statement of need
 
