+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.11] - 2024-04-17 (Current release)
+
+## [1.1.12] - 2024-04-21 (Current release)
+
+- Improvements and enhancements.
+
+## [1.1.11] - 2024-04-17
 
 - Improvements and enhancements.
 
