### 2024-04-22 c9eed08
gauss2dfit performance
