### 2024-04-22 GalagaBits/JWST-FGS-Spelunker#41 @taylorbell57
Everything works great now, except you need to change
```
gaussfit_table = ascii.read('/Users/galagabits/Developer/FGS-Spelunker/JWST-FGS-Spelunker/notebooks/examples/data/pid2079_observation04_visit1_short.dat') # Navigate to the included Gaussian results file
```
to
```
gaussfit_table = ascii.read('./data/pid2079_observation04_visit1_short.dat') # Navigate to the included Gaussian results file
```

You should also remove the cell at the top that has
```
cd /Users/galagabits/Developer/FGS-Spelunker/
```
which isn't relevant for other users
