### 2024-04-22 #6202 @taylorbell57
The paper looks good to me now, and I just pointed out two minor issues with the pixel_centroid_mnemonics.ipynb notebook in Issue https://github.com/GalagaBits/JWST-FGS-Spelunker/issues/41

Once those very minor comments are addressed, I'm happy to recommend the paper for publication @ivastar
