### 2024-04-23 GalagaBits/JWST-FGS-Spelunker#41 @GalagaBits
@taylorbell57, okay! This is fixed
