+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [1.1.14] - 2024-04-22 (Current release)
+## [1.1.14] - 2024-04-23 (Current release)
+
+- Massively improved performance for ``gauss2d_fit``.
+
+## [1.1.14] - 2024-04-22
 
 - Massively improved performance for ``gauss2d_fit``.
 
