### 2024-05-02 GalagaBits/JWST-FGS-Spelunker#26 @evamariaa
Ah I see, that's great thank you! That makes sense then and is good that it is explained here now so that anyone else who runs into this issue also can have a look!
