### 2024-05-02 #6202 @evamariaa
I've gone through the issues now and the remaining things on my check list, as well as re-installed the new version and re-ran notebooks and test. I'm very happy with the improvements!
I would note though that it'd be nice if you could update your example notebooks in the notebooks folder such that they work with your most recent version as this is what everyone will use. For example, when I had version 1.1.16 installed, then the function `spk2.flux_spatial_timelapse_animation` gave me an error and then I realised you were using a different version which you do state in the text at the top of the notebook so totally my bad.
So I don't think this is necessary for this review, but it'll safe a lot of issues/time if you keep those notebooks up-to-date with your current version.

Either way, I'm very happy to recommend the paper for publication. Congrats on this great work!
