### 2024-05-02 #6202 @ivastar
@GalagaBits do you think you can update the notebook as per the suggestion made by @evamariaa? I think this falls under the "Example usage" part of our review. It is important for the examples to reflect the published API.
