### 2024-05-02 #6202 @GalagaBits
@ivastar @evamariaa, yes, I can fix this issue. This turned out to be a bug in the code that prevented animations from working. I am releasing a new version in a few minutes and I will update all notebooks in the folder to use the latest version of `spelunker`.
