+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.17] - 2024-04-23 (Current release)
 
-## [1.1.16] - 2024-04-23 (Current release)
+- Animation functions now work properly.
+
+## [1.1.16] - 2024-04-23
 
 - Massively improved performance for ``gauss2d_fit``.
 
-## [1.1.15] - 2024-04-23 (Current release)
+## [1.1.15] - 2024-04-23
 
 - Massively improved performance for ``gauss2d_fit``.
 
