### 2024-05-02 #6202 @GalagaBits
@ivastar I have updated spelunker to 1.1.17! Is there a way to update the JOSS submission to have the latest version as well?
