### 2024-05-02 #6202 @ivastar
@GalagaBits I will set the final version before publications. But before we get there I have some editorial comments for the manuscript. Please let me know if you have any questions. Ping me when the changes have been done and we will proceed with the post-review checklist.

Overall comments:
- Throughout the text there are several compound words that I think should be written separately. "guide star", "light curve", "time series", "data models" (unless it refers to the `datamodels` library) I think should all be written as separate phrases (including in the title). Happy to be corrected if the JWST documentation for example consistently uses a different style but a brief look through the internet suggests separate words.
- Figure 2 is mentioned before Figure 1 in the text. It is standard that figures should appear in the order in which they are referenced. Please either switch the order of the figures or change the text such that Figure 1 is mentioned first.
- One of my colleagues pointed out that the mnemonics functionality is only available via a package that is currently not pip-installable and (as you point out in the documentation) needs to be separately installed by the user. May I suggest that you document this also in the paper, maybe on line 60 by adding something along the lines of  "This functionality depends on the additional installation of the `jwstuser` library. " Or maybe @nespinoza can push that to pypi....

Smaller comments:
- Line 10 "Through efficient" --> "through an efficient"
- Line 13 "Gaussian parameter and flux time series" --> "Gaussian parameter, flux time series"
- Line 18 "and on the JWST" --> "and in the JWST"
- Line 25 "technique to detect" --> "method to detect"
- Line 33 "uses guidestar to guide" --> "uses guide star to provide information to"
- Line 36 "using this data" --> "using these data"
- Line 43 typo in utilize
- Line 49 "model's orientation" ==> "model orientation"
- Line 60 "abilities" --> "capabilities"
- Line 66-67 "than a guidestar timeseries" --> "than that"
- Line 69 "and spelunker outputs" --> "and its outputs"
- Caption of Figure 1, lines 1-2 "observation 1, and visit 1" --> "observations 1 and visit 1", just remove the comma
- Caption of Figure 1, line 2, typo in "guidance"
- Caption of Figure 1, last line: remove "likely" and maybe change "the functions" to "the core function", as I understand it you are actually demonstrating the that the position of the start is changing during the observations and the ACS
is making corrections to try to keep it constant.
