+++ paper/paper.md
@@ -45,7 +45,7 @@ spk = spelunker.load(pid=1534)
 
 ``spelunker`` uses ``astroquery`` and MAST [@marston_overview_2018] to find and download GS-FG FITS files. There are several functions that can manipulate and analyze guide star data:
 
-- **``gauss2d_fit``** A spatial Gaussian fit is applied to each of the frames loaded into the `spk` object. The Gaussian will fit the amplitude, pixel coordinates, pixel standard deviations, the model orientation, and the background offset. A diagram of Gaussian fitting parameters is shown in \autoref{fig:Gaussian_diagram}. The Gaussian measurements are then stored in ``spk.gaussfit_results`` as an astropy table. Fitting spatial Gaussians to your guide star data can allow you to reveal technical anomalies that might be confused with science data from NIRISS or other JWST instruments and within the Guidestar flux time series (see \autoref{fig:guidestar_1803}).
+- **``gauss2d_fit``** A spatial Gaussian fit is applied to each of the frames loaded into the `spk` object. The Gaussian will fit the amplitude, pixel coordinates, pixel standard deviations, the model orientation, and the background offset. A diagram of Gaussian fitting parameters is shown in \autoref{fig:Gaussian_diagram}. The Gaussian measurements are then stored in ``spk.gaussfit_results`` as an astropy table. Fitting spatial Gaussians to your guide star data can allow you to reveal technical anomalies that might be confused with science data from NIRISS or other JWST instruments and within the guide star flux time series (see \autoref{fig:guidestar_1803}).
 
 
 ![A  snippet from the guide star time series from Cycle 1 GO Program ID 1803, observation 1 and visit 1. **Top** — The guide star time series of PID 1803 after loading it into ``spelunker`` using ``timeseries_binned_plot``. The time series uses the sum of counts in each guide star fine guidance (GS-FG) frame. The data has no significant features. **Middle** — The same time series after applying ``optimize_photometry`` to the guide star light curve. There are now prominent drops in the flux, which were previously unseen with the raw time series data. **Bottom** — Gaussian fitted x pixel coordinate and y pixel coordinate for each frame in this section of time series data. The guide star shifts around in this time series, highlighting the core function of the ACS. \label{fig:guidestar_1803}](timeseries_plot.png)
@@ -54,7 +54,7 @@ spk = spelunker.load(pid=1534)
 
 - **``periodogram``** This function uses the Lomb-Scarle periodogram [@lomb_least-squares_1976;@scargle_studies_1982] to detect periodicities in guide star Gaussian fits. Periods in Gaussian fitted parameters like x and y pixel coordinates highlight systematics for an entire PID.
 
-- **``optimize_photometry``** ``optimize_photometry`` extracts the highest SNR pixels to optimize raw guide star photometry loaded from ``spelunker``. \autoref{fig:guide star_1803} demonstrates that ``optimize_photometry`` reveals more information from guide star time series than that produced by the sum of counts in each frame.
+- **``optimize_photometry``** ``optimize_photometry`` extracts the highest SNR pixels to optimize raw guide star photometry loaded from ``spelunker``. \autoref{fig:guidestar_1803} demonstrates that ``optimize_photometry`` reveals more information from guide star time series than that produced by the sum of counts in each frame.
 
 With the mentioned tools, ``spelunker`` utilizes object oriented programming (OOP) to store handy variables and its outputs, for instance, 1D and 2D time series, guide star time arrays, and JWST data models. Running ``gauss2d_fit``, ``periodogram``, and ``mnemonics`` will store their outputs in accessible attributes. Useful properties of the guide star are stored in these attributes (for instance, guide star galactic coordinates, GAIA ID, and stellar magnitudes). 
 
