### 2024-05-04 #6202 @GalagaBits
How do I archive the release on Zenodo?
