### 2024-05-05 #6202 @GalagaBits
Okay! I finished the additional author tasks!

- Latest version number: v.1.1.17
- Zenodo DOI: 10.5281/zenodo.11115733
