+++ paper/paper.md
@@ -50,7 +50,7 @@ spk = spelunker.load(pid=1534)
 
 ![A  snippet from the guide star time series from Cycle 1 GO Program ID 1803, observation 1 and visit 1. **Top** — The guide star time series of PID 1803 after loading it into ``spelunker`` using ``timeseries_binned_plot``. The time series uses the sum of counts in each guide star fine guidance (GS-FG) frame. The data has no significant features. **Middle** — The same time series after applying ``optimize_photometry`` to the guide star light curve. There are now prominent drops in the flux, which were previously unseen with the raw time series data. **Bottom** — Gaussian fitted x pixel coordinate and y pixel coordinate for each frame in this section of time series data. The guide star shifts around in this time series, highlighting the core function of the ACS. \label{fig:guidestar_1803}](timeseries_plot.png)
 
-- ** ``mnemonics``** Users can access JWST engineering telemetry and mnemonics using the ``mnemonics`` function. With a MAST API token, any mnemonic is accessible. High-gain antenna (HGA) movement and NIRISS filter wheel current are two examples of events that overplot science data to identify technical events on the telescope. Anomalies detection in guide star data or data from NIRISS, NIRCAM, NIRSpec, and MIRI is one of the primary capabilities of this function. Using ``mnemonics`` requires the additional installation of the ``jwstuser`` library.
+- **``mnemonics``** Users can access JWST engineering telemetry and mnemonics using the ``mnemonics`` function. With a MAST API token, any mnemonic is accessible. High-gain antenna (HGA) movement and NIRISS filter wheel current are two examples of events that overplot science data to identify technical events on the telescope. Anomalies detection in guide star data or data from NIRISS, NIRCAM, NIRSpec, and MIRI is one of the primary capabilities of this function. Using ``mnemonics`` requires the additional installation of the ``jwstuser`` library.
 
 - **``periodogram``** This function uses the Lomb-Scarle periodogram [@lomb_least-squares_1976;@scargle_studies_1982] to detect periodicities in guide star Gaussian fits. Periods in Gaussian fitted parameters like x and y pixel coordinates highlight systematics for an entire PID.
 
