### 2024-05-07 #6202 @ivastar
Yes, I recommended publication as the editor for this submission, @dfm is the editor in chief for the AASS track and he does the honors for the final publication.
