### 2024-05-08 #6202 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
